[Split CSSPropertyParserHelpers] Part 1: Primitives, Lists, and Color https://bugs.webkit.org/show_bug.cgi?id=272774
Reviewed by Tim Nguyen.

CSSPropertyParserHelpers.cpp is enormous at 8872 lines long, which
makes it hard to work with and makes compile/fix cycles unwieldily.

This starts the process of breaking it up into logical chunks, starting
with the the primitive consumers that correspond to token types (number,
ident, percent, etc.), lists and color (color weighs in at ~1600 lines
all by itself).

No code changes are being made here, this is purely moving things around.

CSSPropertyParserHelpers.cpp goes from 8872 lines to 5776 lines, that's
3096 lines moved.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/ShorthandSerializer.cpp:
* Source/WebCore/css/parser/CSSCalcParser.cpp: Added.
* Source/WebCore/css/parser/CSSCalcParser.h: Added.
* Source/WebCore/css/parser/CSSParserImpl.cpp:
* Source/WebCore/css/parser/CSSPropertyParser.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Angle.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Angle.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Color.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Color.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Ident.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Ident.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Integer.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Integer.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Length.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Length.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+List.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Meta.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+None.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+None.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Number.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Number.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Percent.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Percent.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Primitives.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Primitives.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Resolution.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Resolution.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Time.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Time.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
* Source/WebCore/css/parser/CSSPropertyParserHelpers.h:
* Source/WebCore/css/parser/CSSPropertyParserWorkerSafe.cpp:
* Source/WebCore/css/process-css-properties.py:
* Source/WebCore/css/query/GenericMediaQueryParser.cpp:
* Source/WebCore/page/IntersectionObserver.cpp:
* Source/WebCore/style/StyleBuilderState.h:

Canonical link: https://commits.webkit.org/277585@main
