+++ Source/WebCore/css/parser/CSSParserImpl.cpp
@@ -43,6 +43,9 @@
 #include "CSSParserObserver.h"
 #include "CSSParserObserverWrapper.h"
 #include "CSSPropertyParser.h"
+#include "CSSPropertyParserConsumer+Ident.h"
+#include "CSSPropertyParserConsumer+Integer.h"
+#include "CSSPropertyParserConsumer+Length.h"
 #include "CSSSelectorList.h"
 #include "CSSSelectorParser.h"
 #include "CSSStyleSheet.h"
