+++ Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp
@@ -30,6 +30,19 @@
 #include "config.h"
 #include "CSSPropertyParserHelpers.h"
 #include "CSSPropertyParser.h"
+#include "CSSPropertyParserConsumer+Angle.h"
+#include "CSSPropertyParserConsumer+Color.h"
+#include "CSSPropertyParserConsumer+Ident.h"
+#include "CSSPropertyParserConsumer+Integer.h"
+#include "CSSPropertyParserConsumer+Length.h"
+#include "CSSPropertyParserConsumer+List.h"
+#include "CSSPropertyParserConsumer+Meta.h"
+#include "CSSPropertyParserConsumer+None.h"
+#include "CSSPropertyParserConsumer+Number.h"
+#include "CSSPropertyParserConsumer+Percent.h"
+#include "CSSPropertyParserConsumer+Primitives.h"
+#include "CSSPropertyParserConsumer+Resolution.h"
+#include "CSSPropertyParserConsumer+Time.h"
 #include "CSSPropertyParsing.h"
 
 #include "CSSBackgroundRepeatValue.h"
@@ -77,7 +90,6 @@
 #include "CSSSubgridValue.h"
 #include "CSSTimingFunctionValue.h"
 #include "CSSTransformListValue.h"
-#include "CSSUnresolvedColor.h"
 #include "CSSValueKeywords.h"
 #include "CSSValuePair.h"
 #include "CSSValuePool.h"
@@ -85,10 +97,7 @@
 #include "CSSVariableParser.h"
 #include "CSSViewValue.h"
 #include "CalculationCategory.h"
-#include "ColorConversion.h"
 #include "ColorInterpolation.h"
-#include "ColorLuminance.h"
-#include "ColorNormalization.h"
 #include "FontCustomPlatformData.h"
 #include "FontFace.h"
 #include "Logging.h"
@@ -107,742 +116,7 @@ namespace WebCore {
 
 namespace CSSPropertyParserHelpers {
 
-using NumberOrPercentRaw = std::variant<NumberRaw, PercentRaw>;
-using NumberOrNoneRaw = std::variant<NumberRaw, NoneRaw>;
-using PercentOrNoneRaw = std::variant<PercentRaw, NoneRaw>;
-using NumberOrPercentOrNoneRaw = std::variant<NumberRaw, PercentRaw, NoneRaw>;
-using AngleOrNumberRaw = std::variant<AngleRaw, NumberRaw>;
-using AngleOrNumberOrNoneRaw = std::variant<AngleRaw, NumberRaw, NoneRaw>;
-using ColorOrUnresolvedColor = std::variant<Color, CSSUnresolvedColor>;
-
-bool consumeCommaIncludingWhitespace(CSSParserTokenRange& range)
-{
-    CSSParserToken value = range.peek();
-    if (value.type() != CommaToken)
-        return false;
-    range.consumeIncludingWhitespace();
-    return true;
-}
-
-bool consumeSlashIncludingWhitespace(CSSParserTokenRange& range)
-{
-    CSSParserToken value = range.peek();
-    if (value.type() != DelimiterToken || value.delimiter() != '/')
-        return false;
-    range.consumeIncludingWhitespace();
-    return true;
-}
-
-CSSParserTokenRange consumeFunction(CSSParserTokenRange& range)
-{
-    ASSERT(range.peek().type() == FunctionToken);
-    CSSParserTokenRange contents = range.consumeBlock();
-    range.consumeWhitespace();
-    contents.consumeWhitespace();
-    return contents;
-}
-
-template<typename Map>
-static std::optional<typename Map::ValueType> consumeIdentUsingMapping(CSSParserTokenRange& range, Map& map)
-{
-    if (auto value = map.tryGet(range.peek().id())) {
-        range.consumeIncludingWhitespace();
-        return std::make_optional(*value);
-    }
-    return std::nullopt;
-}
-
-template<typename Map>
-static std::optional<typename Map::ValueType> peekIdentUsingMapping(CSSParserTokenRange& range, Map& map)
-{
-    if (auto value = map.tryGet(range.peek().id()))
-        return std::make_optional(*value);
-    return std::nullopt;
-}
-
-inline bool shouldAcceptUnitlessValue(double value, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-{
-    // FIXME: Presentational HTML attributes shouldn't use the CSS parser for lengths.
-
-    if (!value && unitlessZero == UnitlessZeroQuirk::Allow)
-        return true;
-
-    if (isUnitlessValueParsingEnabledForMode(parserMode))
-        return true;
-
-    return parserMode == HTMLQuirksMode && unitless == UnitlessQuirk::Allow;
-}
-
-static bool canConsumeCalcValue(CalculationCategory category, CSSParserMode parserMode)
-{
-    if (category == CalculationCategory::Length || category == CalculationCategory::Percent || category == CalculationCategory::PercentLength)
-        return true;
-
-    if (parserMode != SVGAttributeMode)
-        return false;
-
-    if (category == CalculationCategory::Number || category == CalculationCategory::PercentNumber)
-        return true;
-
-    return false;
-}
-
-// FIXME: consider pulling in the parsing logic from CSSCalcExpressionNodeParser.
-class CalcParser {
-public:
-    explicit CalcParser(CSSParserTokenRange& range, CalculationCategory destinationCategory, ValueRange valueRange = ValueRange::All, const CSSCalcSymbolTable& symbolTable = { }, NegativePercentagePolicy negativePercentagePolicy = NegativePercentagePolicy::Forbid)
-        : m_sourceRange(range)
-        , m_range(range)
-    {
-        auto functionId = range.peek().functionId();
-        if (CSSCalcValue::isCalcFunction(functionId))
-            m_value = CSSCalcValue::create(functionId, consumeFunction(m_range), destinationCategory, valueRange, symbolTable, negativePercentagePolicy == NegativePercentagePolicy::Allow);
-    }
-
-    const CSSCalcValue* value() const { return m_value.get(); }
-
-    RefPtr<CSSPrimitiveValue> consumeValue()
-    {
-        if (!m_value)
-            return nullptr;
-        m_sourceRange = m_range;
-        return CSSPrimitiveValue::create(m_value.releaseNonNull());
-    }
-
-    RefPtr<CSSPrimitiveValue> consumeValueIfCategory(CalculationCategory category)
-    {
-        if (m_value && m_value->category() != category) {
-            LOG_WITH_STREAM(Calc, stream << "CalcParser::consumeValueIfCategory - failing because calc category " << m_value->category() << " does not match requested category " << category);
-            return nullptr;
-        }
-        return consumeValue();
-    }
-
-private:
-    CSSParserTokenRange& m_sourceRange;
-    CSSParserTokenRange m_range;
-    RefPtr<CSSCalcValue> m_value;
-};
-
-// MARK: - Primitive value consumers for callers that know the token type.
-
-static RefPtr<CSSCalcValue> consumeCalcRawWithKnownTokenTypeFunction(CSSParserTokenRange& range, CalculationCategory category, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange)
-{
-    ASSERT(range.peek().type() == FunctionToken);
-
-    auto functionId = range.peek().functionId();
-    if (!CSSCalcValue::isCalcFunction(functionId))
-        return nullptr;
-
-    RefPtr calcValue = CSSCalcValue::create(functionId, consumeFunction(range), category, valueRange, symbolTable);
-    if (calcValue && calcValue->category() == category)
-        return calcValue;
-
-    return nullptr;
-}
-
-// MARK: Integer (Raw)
-
-static constexpr double computeMinimumValue(IntegerValueRange range)
-{
-    switch (range) {
-    case IntegerValueRange::All:
-        return -std::numeric_limits<double>::infinity();
-    case IntegerValueRange::NonNegative:
-        return 0.0;
-    case IntegerValueRange::Positive:
-        return 1.0;
-    }
-
-    RELEASE_ASSERT_NOT_REACHED_UNDER_CONSTEXPR_CONTEXT();
-
-    return 0.0;
-}
-
-// MARK: Integer (Raw)
-
-template<typename IntType, IntegerValueRange integerRange>
-struct IntegerTypeRawKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static std::optional<IntType> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        auto rangeCopy = range;
-        if (RefPtr value = consumeCalcRawWithKnownTokenTypeFunction(rangeCopy, CalculationCategory::Number, { }, ValueRange::All)) {
-            range = rangeCopy;
-            // https://drafts.csswg.org/css-values-4/#integers
-            // Rounding to the nearest integer requires rounding in the direction of +∞ when the fractional portion is exactly 0.5.
-            return clampTo<IntType>(std::floor(std::max(value->doubleValue(), computeMinimumValue(integerRange)) + 0.5));
-        }
-
-        return std::nullopt;
-    }
-};
-
-template<typename IntType, IntegerValueRange integerRange>
-struct IntegerTypeRawKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static std::optional<IntType> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        if (range.peek().numericValueType() == NumberValueType || range.peek().numericValue() < computeMinimumValue(integerRange))
-            return std::nullopt;
-        return clampTo<IntType>(range.consumeIncludingWhitespace().numericValue());
-    }
-};
-
-// MARK: Integer (CSSPrimitiveValue - maintaining calc)
-
-template<typename IntType, IntegerValueRange integerRange>
-struct IntegerTypeKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        if (auto integer = IntegerTypeRawKnownTokenTypeFunctionConsumer<IntType, integerRange>::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero))
-            return CSSPrimitiveValue::createInteger(*integer);
-        return nullptr;
-    }
-};
-
-template<typename IntType, IntegerValueRange integerRange>
-struct IntegerTypeKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        if (auto integer = IntegerTypeRawKnownTokenTypeNumberConsumer<IntType, integerRange>::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero))
-            return CSSPrimitiveValue::createInteger(*integer);
-        return nullptr;
-    }
-};
-
-// MARK: Number (Raw)
-
-static std::optional<NumberRaw> validatedNumberRaw(double value, ValueRange valueRange)
-{
-    if (valueRange == ValueRange::NonNegative && value < 0)
-        return std::nullopt;
-    return {{ value }};
-}
-
-struct NumberRawKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static std::optional<NumberRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        auto rangeCopy = range;
-        if (RefPtr value = consumeCalcRawWithKnownTokenTypeFunction(rangeCopy, CalculationCategory::Number, symbolTable, valueRange)) {
-            if (auto validatedValue = validatedNumberRaw(value->doubleValue(), valueRange)) {
-                range = rangeCopy;
-                return validatedValue;
-            }
-        }
-
-        return std::nullopt;
-    }
-};
-
-struct NumberRawKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static std::optional<NumberRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        if (auto validatedValue = validatedNumberRaw(range.peek().numericValue(), valueRange)) {
-            range.consumeIncludingWhitespace();
-            return validatedValue;
-        }
-        return std::nullopt;
-    }
-};
-
-struct NumberRawKnownTokenTypeIdentConsumer {
-    static constexpr CSSParserTokenType tokenType = IdentToken;
-    static std::optional<NumberRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == IdentToken);
-
-        if (auto variable = symbolTable.get(range.peek().id())) {
-            switch (variable->type) {
-            case CSSUnitType::CSS_NUMBER:
-                if (auto validatedValue = validatedNumberRaw(variable->value, valueRange)) {
-                    range.consumeIncludingWhitespace();
-                    return validatedValue;
-                }
-                break;
-
-            default:
-                break;
-            }
-        }
-        return std::nullopt;
-    }
-};
-
-// MARK: Number (CSSPrimitiveValue - maintaining calc)
-
-struct NumberCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        CalcParser parser(range, CalculationCategory::Number, valueRange, symbolTable);
-        return parser.consumeValueIfCategory(CalculationCategory::Number);
-    }
-};
-
-struct NumberCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        auto token = range.peek();
-
-        if (auto validatedValue = validatedNumberRaw(token.numericValue(), valueRange)) {
-            auto unitType = token.unitType();
-            range.consumeIncludingWhitespace();
-            return CSSPrimitiveValue::create(validatedValue->value, unitType);
-        }
-        return nullptr;
-    }
-};
-
-// MARK: Percent (raw)
-
-static std::optional<PercentRaw> validatedPercentRaw(double value, ValueRange valueRange)
-{
-    if (valueRange == ValueRange::NonNegative && value < 0)
-        return std::nullopt;
-    if (std::isinf(value))
-        return std::nullopt;
-    return {{ value }};
-}
-
-struct PercentRawKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static std::optional<PercentRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        auto rangeCopy = range;
-        if (RefPtr value = consumeCalcRawWithKnownTokenTypeFunction(rangeCopy, CalculationCategory::Percent, symbolTable, valueRange)) {
-            range = rangeCopy;
-
-            // FIXME: Should this validate the calc value as is done for the NumberRaw variant?
-            return {{ value->doubleValue() }};
-        }
-
-        return std::nullopt;
-    }
-};
-
-struct PercentRawKnownTokenTypePercentConsumer {
-    static constexpr CSSParserTokenType tokenType = PercentageToken;
-    static std::optional<PercentRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == PercentageToken);
-
-        if (auto validatedValue = validatedPercentRaw(range.peek().numericValue(), valueRange)) {
-            range.consumeIncludingWhitespace();
-            return validatedValue;
-        }
-        return std::nullopt;
-    }
-};
-
-struct PercentRawKnownTokenTypeIdentConsumer {
-    static constexpr CSSParserTokenType tokenType = IdentToken;
-    static std::optional<PercentRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == IdentToken);
-
-        if (auto variable = symbolTable.get(range.peek().id())) {
-            switch (variable->type) {
-            case CSSUnitType::CSS_PERCENTAGE:
-                if (auto validatedValue = validatedPercentRaw(variable->value, valueRange)) {
-                    range.consumeIncludingWhitespace();
-                    return validatedValue;
-                }
-                break;
-
-            default:
-                break;
-            }
-        }
-        return std::nullopt;
-    }
-};
-
-// MARK: Percent (CSSPrimitiveValue - maintaining calc)
-
-struct PercentCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        CalcParser parser(range, CalculationCategory::Percent, valueRange, symbolTable);
-        return parser.consumeValueIfCategory(CalculationCategory::Percent);
-    }
-};
-
-struct PercentCSSPrimitiveValueWithCalcWithKnownTokenTypePercentConsumer {
-    static constexpr CSSParserTokenType tokenType = PercentageToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == PercentageToken);
-
-        if (auto validatedValue = validatedPercentRaw(range.peek().numericValue(), valueRange)) {
-            range.consumeIncludingWhitespace();
-            return CSSPrimitiveValue::create(validatedValue->value, CSSUnitType::CSS_PERCENTAGE);
-        }
-        return nullptr;
-    }
-};
-
-// MARK: Length (raw)
-
-static std::optional<double> validatedLengthRaw(double value, ValueRange valueRange)
-{
-    if (valueRange == ValueRange::NonNegative && value < 0)
-        return std::nullopt;
-    if (std::isinf(value))
-        return std::nullopt;
-    return value;
-}
-
-struct LengthRawKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static std::optional<LengthRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        auto rangeCopy = range;
-        if (RefPtr value = consumeCalcRawWithKnownTokenTypeFunction(rangeCopy, CalculationCategory::Length, symbolTable, valueRange)) {
-            range = rangeCopy;
-
-            // FIXME: Should this validate the calc value as is done for the NumberRaw variant?
-            return { { value->primitiveType(), value->doubleValue() } };
-        }
-        return std::nullopt;
-    }
-};
-
-struct LengthRawKnownTokenTypeDimensionConsumer {
-    static constexpr CSSParserTokenType tokenType = DimensionToken;
-    static std::optional<LengthRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == DimensionToken);
-
-        auto& token = range.peek();
-
-        auto unitType = token.unitType();
-        switch (unitType) {
-        case CSSUnitType::CSS_QUIRKY_EM:
-            if (!isUASheetBehavior(parserMode))
-                return std::nullopt;
-            FALLTHROUGH;
-        case CSSUnitType::CSS_EM:
-        case CSSUnitType::CSS_REM:
-        case CSSUnitType::CSS_LH:
-        case CSSUnitType::CSS_RLH:
-        case CSSUnitType::CSS_CAP:
-        case CSSUnitType::CSS_RCAP:
-        case CSSUnitType::CSS_CH:
-        case CSSUnitType::CSS_RCH:
-        case CSSUnitType::CSS_IC:
-        case CSSUnitType::CSS_RIC:
-        case CSSUnitType::CSS_EX:
-        case CSSUnitType::CSS_REX:
-        case CSSUnitType::CSS_PX:
-        case CSSUnitType::CSS_CM:
-        case CSSUnitType::CSS_MM:
-        case CSSUnitType::CSS_IN:
-        case CSSUnitType::CSS_PT:
-        case CSSUnitType::CSS_PC:
-        case CSSUnitType::CSS_VW:
-        case CSSUnitType::CSS_VH:
-        case CSSUnitType::CSS_VMIN:
-        case CSSUnitType::CSS_VMAX:
-        case CSSUnitType::CSS_VB:
-        case CSSUnitType::CSS_VI:
-        case CSSUnitType::CSS_SVW:
-        case CSSUnitType::CSS_SVH:
-        case CSSUnitType::CSS_SVMIN:
-        case CSSUnitType::CSS_SVMAX:
-        case CSSUnitType::CSS_SVB:
-        case CSSUnitType::CSS_SVI:
-        case CSSUnitType::CSS_LVW:
-        case CSSUnitType::CSS_LVH:
-        case CSSUnitType::CSS_LVMIN:
-        case CSSUnitType::CSS_LVMAX:
-        case CSSUnitType::CSS_LVB:
-        case CSSUnitType::CSS_LVI:
-        case CSSUnitType::CSS_DVW:
-        case CSSUnitType::CSS_DVH:
-        case CSSUnitType::CSS_DVMIN:
-        case CSSUnitType::CSS_DVMAX:
-        case CSSUnitType::CSS_DVB:
-        case CSSUnitType::CSS_DVI:
-        case CSSUnitType::CSS_Q:
-        case CSSUnitType::CSS_CQW:
-        case CSSUnitType::CSS_CQH:
-        case CSSUnitType::CSS_CQI:
-        case CSSUnitType::CSS_CQB:
-        case CSSUnitType::CSS_CQMIN:
-        case CSSUnitType::CSS_CQMAX:
-            break;
-        default:
-            return std::nullopt;
-        }
-
-        if (auto validatedValue = validatedLengthRaw(token.numericValue(), valueRange)) {
-            range.consumeIncludingWhitespace();
-            return { { unitType, *validatedValue } };
-        }
-        return std::nullopt;
-    }
-};
-
-struct LengthRawKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static std::optional<LengthRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        auto& token = range.peek();
-
-        if (!shouldAcceptUnitlessValue(token.numericValue(), parserMode, unitless, unitlessZero))
-            return std::nullopt;
-
-        if (auto validatedValue = validatedLengthRaw(token.numericValue(), valueRange)) {
-            range.consumeIncludingWhitespace();
-            return { { CSSUnitType::CSS_PX, *validatedValue } };
-        }
-        return std::nullopt;
-    }
-};
-
-// MARK: Length (CSSPrimitiveValue - maintaining calc)
-
-struct LengthCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        CalcParser parser(range, CalculationCategory::Length, valueRange, symbolTable);
-        return parser.consumeValueIfCategory(CalculationCategory::Length);
-    }
-};
-
-struct LengthCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer {
-    static constexpr CSSParserTokenType tokenType = DimensionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-    {
-        ASSERT(range.peek().type() == DimensionToken);
-
-        if (auto lengthRaw = LengthRawKnownTokenTypeDimensionConsumer::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero))
-            return CSSPrimitiveValue::create(lengthRaw->value, lengthRaw->type);
-        return nullptr;
-    }
-};
-
-struct LengthCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        if (auto lengthRaw = LengthRawKnownTokenTypeNumberConsumer::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero))
-            return CSSPrimitiveValue::create(lengthRaw->value, lengthRaw->type);
-        return nullptr;
-    }
-};
-
-// MARK: Angle (raw)
-
-struct AngleRawKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static std::optional<AngleRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        auto rangeCopy = range;
-        if (RefPtr value = consumeCalcRawWithKnownTokenTypeFunction(rangeCopy, CalculationCategory::Angle, symbolTable, valueRange)) {
-            range = rangeCopy;
-            return { { value->primitiveType(), value->doubleValue() } };
-        }
-        return std::nullopt;
-    }
-};
-
-struct AngleRawKnownTokenTypeDimensionConsumer {
-    static constexpr CSSParserTokenType tokenType = DimensionToken;
-    static std::optional<AngleRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == DimensionToken);
-
-        auto unitType = range.peek().unitType();
-        switch (unitType) {
-        case CSSUnitType::CSS_DEG:
-        case CSSUnitType::CSS_RAD:
-        case CSSUnitType::CSS_GRAD:
-        case CSSUnitType::CSS_TURN:
-            return { { unitType, range.consumeIncludingWhitespace().numericValue() } };
-        default:
-            break;
-        }
-
-        return std::nullopt;
-    }
-};
-
-struct AngleRawKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static std::optional<AngleRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        if (shouldAcceptUnitlessValue(range.peek().numericValue(), parserMode, unitless, unitlessZero))
-            return { { CSSUnitType::CSS_DEG, range.consumeIncludingWhitespace().numericValue() } };
-        return std::nullopt;
-    }
-};
-
-// MARK: Angle (CSSPrimitiveValue - maintaining calc)
-
-struct AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        CalcParser parser(range, CalculationCategory::Angle, valueRange, symbolTable);
-        return parser.consumeValueIfCategory(CalculationCategory::Angle);
-    }
-};
-
-struct AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer {
-    static constexpr CSSParserTokenType tokenType = DimensionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-    {
-        ASSERT(range.peek().type() == DimensionToken);
-
-        if (auto angleRaw = AngleRawKnownTokenTypeDimensionConsumer::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero))
-            return CSSPrimitiveValue::create(angleRaw->value, angleRaw->type);
-        return nullptr;
-    }
-};
-
-struct AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        if (auto angleRaw = AngleRawKnownTokenTypeNumberConsumer::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero))
-            return CSSPrimitiveValue::create(angleRaw->value, angleRaw->type);
-        return nullptr;
-    }
-};
-
-// MARK: Time (CSSPrimitiveValue - maintaining calc)
-
-struct TimeCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        CalcParser parser(range, CalculationCategory::Time, valueRange, symbolTable);
-        return parser.consumeValueIfCategory(CalculationCategory::Time);
-    }
-};
-
-struct TimeCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer {
-    static constexpr CSSParserTokenType tokenType = DimensionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == DimensionToken);
-
-        if (valueRange == ValueRange::NonNegative && range.peek().numericValue() < 0)
-            return nullptr;
-
-        if (auto unit = range.peek().unitType(); unit == CSSUnitType::CSS_MS || unit == CSSUnitType::CSS_S)
-            return CSSPrimitiveValue::create(range.consumeIncludingWhitespace().numericValue(), unit);
-
-        return nullptr;
-    }
-};
-
-struct TimeCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer {
-    static constexpr CSSParserTokenType tokenType = NumberToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == NumberToken);
-
-        if (unitless == UnitlessQuirk::Allow && shouldAcceptUnitlessValue(range.peek().numericValue(), parserMode, unitless, UnitlessZeroQuirk::Allow)) {
-            if (valueRange == ValueRange::NonNegative && range.peek().numericValue() < 0)
-                return nullptr;
-            return CSSPrimitiveValue::create(range.consumeIncludingWhitespace().numericValue(), CSSUnitType::CSS_MS);
-        }
-        return nullptr;
-    }
-};
-
-// MARK: Resolution (CSSPrimitiveValue - maintaining calc)
-
-struct ResolutionCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer {
-    static constexpr CSSParserTokenType tokenType = FunctionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == FunctionToken);
-
-        CalcParser parser(range, CalculationCategory::Resolution, valueRange, symbolTable);
-        return parser.consumeValueIfCategory(CalculationCategory::Resolution);
-    }
-};
-
-struct ResolutionCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer {
-    static constexpr CSSParserTokenType tokenType = DimensionToken;
-    static RefPtr<CSSPrimitiveValue> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == DimensionToken);
-
-        if (auto unit = range.peek().unitType(); unit == CSSUnitType::CSS_DPPX || unit == CSSUnitType::CSS_X || unit == CSSUnitType::CSS_DPI || unit == CSSUnitType::CSS_DPCM) {
-            if (valueRange == ValueRange::NonNegative && range.peek().numericValue() < 0)
-                return nullptr;
-            return CSSPrimitiveValue::create(range.consumeIncludingWhitespace().numericValue(), unit);
-        }
-
-        return nullptr;
-    }
-};
-
-// MARK: None (Raw)
-
-struct NoneRawKnownTokenTypeIdentConsumer {
-    static constexpr CSSParserTokenType tokenType = IdentToken;
-    static std::optional<NoneRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable&, ValueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == IdentToken);
-
-        if (range.peek().id() == CSSValueNone) {
-            range.consumeIncludingWhitespace();
-            return NoneRaw { };
-        }
-
-        return std::nullopt;
-    }
-};
-
+// MARK: - Non-primitive consumers.
 
 // MARK: Image Set Type
 
@@ -867,538 +141,6 @@ struct ImageSetTypeCSSPrimitiveValueKnownTokenTypeFunctionConsumer {
     }
 };
 
-// MARK: Specialized combination consumers.
-
-// FIXME: It would be good to find a way to synthesize this from an angle and number specific variants.
-struct AngleOrNumberRawKnownTokenTypeIdentConsumer {
-    static constexpr CSSParserTokenType tokenType = IdentToken;
-    static std::optional<AngleOrNumberRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == IdentToken);
-
-        if (auto variable = symbolTable.get(range.peek().id())) {
-            switch (variable->type) {
-            case CSSUnitType::CSS_DEG:
-            case CSSUnitType::CSS_RAD:
-            case CSSUnitType::CSS_GRAD:
-            case CSSUnitType::CSS_TURN:
-                range.consumeIncludingWhitespace();
-                return AngleRaw { variable->type, variable->value };
-
-            case CSSUnitType::CSS_NUMBER:
-                range.consumeIncludingWhitespace();
-                return NumberRaw { variable->value };
-
-            default:
-                break;
-            }
-        }
-
-        return std::nullopt;
-    }
-};
-
-// FIXME: It would be good to find a way to synthesize this from an number and percent specific variants.
-struct NumberOrPercentRawKnownTokenTypeIdentConsumer {
-    static constexpr CSSParserTokenType tokenType = IdentToken;
-    static std::optional<NumberOrPercentRaw> consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode, UnitlessQuirk, UnitlessZeroQuirk)
-    {
-        ASSERT(range.peek().type() == IdentToken);
-
-        if (auto variable = symbolTable.get(range.peek().id())) {
-            switch (variable->type) {
-            case CSSUnitType::CSS_PERCENTAGE:
-                if (auto validatedValue = validatedPercentRaw(variable->value, valueRange)) {
-                    range.consumeIncludingWhitespace();
-                    return {{ *validatedValue }};
-                }
-                break;
-
-            case CSSUnitType::CSS_NUMBER:
-                if (auto validatedValue = validatedNumberRaw(variable->value, valueRange)) {
-                    range.consumeIncludingWhitespace();
-                    return {{ *validatedValue }};
-                }
-                break;
-
-            default:
-                break;
-            }
-        }
-
-        return std::nullopt;
-    }
-};
-
-// MARK: - Meta Consumers
-
-template<typename Consumer, typename = void>
-struct TransformApplier {
-    template<typename T> static decltype(auto) apply(T value) { return value; }
-};
-
-template<typename Consumer>
-struct TransformApplier<Consumer, typename std::void_t<typename Consumer::Transformer>> {
-    template<typename T> static decltype(auto) apply(T value) { return Consumer::Transformer::transform(value); }
-};
-
-template<typename Consumer, typename T>
-static decltype(auto) applyTransform(T value)
-{
-    return TransformApplier<Consumer>::apply(value);
-}
-
-// Transformers:
-
-template<typename ResultType>
-struct IdentityTransformer {
-    using Result = ResultType;
-    static Result transform(Result value) { return value; }
-};
-
-template<typename ResultType>
-struct RawIdentityTransformer {
-    using Result = std::optional<ResultType>;
-
-    static Result transform(Result value) { return value; }
-
-    template<typename... Types>
-    static Result transform(const std::variant<Types...>& value)
-    {
-        return WTF::switchOn(value, [] (auto value) -> Result { return value; });
-    }
-
-    template<typename T>
-    static Result transform(std::optional<T> value)
-    {
-        if (value)
-            return transform(*value);
-        return std::nullopt;
-    }
-};
-
-template<typename Parent, typename ResultType>
-struct RawVariantTransformerBase {
-    using Result = std::optional<ResultType>;
-
-    template<typename... Types>
-    static Result transform(const std::variant<Types...>& value)
-    {
-        return WTF::switchOn(value, [] (auto value) { return Parent::transform(value); });
-    }
-
-    static typename Result::value_type transform(typename Result::value_type value)
-    {
-        return value;
-    }
-
-    template<typename T>
-    static Result transform(std::optional<T> value)
-    {
-        if (value)
-            return Parent::transform(*value);
-        return std::nullopt;
-    }
-};
-
-struct NumberOrPercentDividedBy100Transformer : RawVariantTransformerBase<NumberOrPercentDividedBy100Transformer, double> {
-    using RawVariantTransformerBase<NumberOrPercentDividedBy100Transformer, double>::transform;
-
-    static double transform(NumberRaw value)
-    {
-        return value.value;
-    }
-
-    static double transform(PercentRaw value)
-    {
-        return value.value / 100.0;
-    }
-};
-
-// MARK: MetaConsumerDispatcher
-
-template<CSSParserTokenType tokenType, typename Consumer, typename = void>
-struct MetaConsumerDispatcher {
-    template<typename... Args>
-    static typename Consumer::Result consume(Args&&...)
-    {
-        return { };
-    }
-};
-
-template<typename Consumer>
-struct MetaConsumerDispatcher<FunctionToken, Consumer, typename std::void_t<typename Consumer::FunctionToken>> {
-    template<typename... Args>
-    static typename Consumer::Result consume(Args&&... args)
-    {
-        return applyTransform<Consumer>(Consumer::FunctionToken::consume(std::forward<Args>(args)...));
-    }
-};
-
-template<typename Consumer>
-struct MetaConsumerDispatcher<NumberToken, Consumer, typename std::void_t<typename Consumer::NumberToken>> {
-    template<typename... Args>
-    static typename Consumer::Result consume(Args&&... args)
-    {
-        return applyTransform<Consumer>(Consumer::NumberToken::consume(std::forward<Args>(args)...));
-    }
-};
-
-template<typename Consumer>
-struct MetaConsumerDispatcher<PercentageToken, Consumer, typename std::void_t<typename Consumer::PercentageToken>> {
-    template<typename... Args>
-    static typename Consumer::Result consume(Args&&... args)
-    {
-        return applyTransform<Consumer>(Consumer::PercentageToken::consume(std::forward<Args>(args)...));
-    }
-};
-
-template<typename Consumer>
-struct MetaConsumerDispatcher<DimensionToken, Consumer, typename std::void_t<typename Consumer::DimensionToken>> {
-    template<typename... Args>
-    static typename Consumer::Result consume(Args&&... args)
-    {
-        return applyTransform<Consumer>(Consumer::DimensionToken::consume(std::forward<Args>(args)...));
-    }
-};
-
-template<typename Consumer>
-struct MetaConsumerDispatcher<IdentToken, Consumer, typename std::void_t<typename Consumer::IdentToken>> {
-    template<typename... Args>
-    static typename Consumer::Result consume(Args&&... args)
-    {
-        return applyTransform<Consumer>(Consumer::IdentToken::consume(std::forward<Args>(args)...));
-    }
-};
-
-template<typename Consumer, typename... Args>
-auto consumeMetaConsumer(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero, Args&&... args) -> typename Consumer::Result
-{
-    switch (range.peek().type()) {
-    case FunctionToken:
-        return MetaConsumerDispatcher<FunctionToken, Consumer>::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero, std::forward<Args>(args)...);
-
-    case NumberToken:
-        return MetaConsumerDispatcher<NumberToken, Consumer>::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero, std::forward<Args>(args)...);
-
-    case PercentageToken:
-        return MetaConsumerDispatcher<PercentageToken, Consumer>::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero, std::forward<Args>(args)...);
-
-    case DimensionToken:
-        return MetaConsumerDispatcher<DimensionToken, Consumer>::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero, std::forward<Args>(args)...);
-
-    case IdentToken:
-        return MetaConsumerDispatcher<IdentToken, Consumer>::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero, std::forward<Args>(args)...);
-
-    default:
-        return { };
-    }
-}
-
-// MARK: SameTokenMetaConsumer
-
-template<typename Transformer, typename ConsumersTuple, size_t I>
-struct SameTokenMetaConsumerApplier {
-    template<typename... Args>
-    static typename Transformer::Result consume(Args&&... args)
-    {
-        using SelectedConsumer = std::tuple_element_t<I - 1, ConsumersTuple>;
-        if (auto result = Transformer::transform(SelectedConsumer::consume(std::forward<Args>(args)...)))
-            return result;
-        return SameTokenMetaConsumerApplier<Transformer, ConsumersTuple, I - 1>::consume(std::forward<Args>(args)...);
-    }
-};
-
-template<typename Transformer, typename ConsumersTuple>
-struct SameTokenMetaConsumerApplier<Transformer, ConsumersTuple, 0> {
-    template<typename... Args>
-    static typename Transformer::Result consume(Args&&...)
-    {
-        return typename Transformer::Result { };
-    }
-};
-
-template<typename Transformer, typename T, typename... Ts>
-struct SameTokenMetaConsumer {
-    static_assert(std::conjunction_v<std::bool_constant<Ts::tokenType == T::tokenType>...>, "All Consumers passed to SameTokenMetaConsumer must have the same tokenType");
-    using ConsumersTuple = std::tuple<T, Ts...>;
-
-    static constexpr CSSParserTokenType tokenType = T::tokenType;
-    template<typename... Args>
-    static typename Transformer::Result consume(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero, Args&&... args)
-    {
-        ASSERT(range.peek().type() == tokenType);
-        return SameTokenMetaConsumerApplier<Transformer, ConsumersTuple, std::tuple_size_v<ConsumersTuple>>::consume(range, symbolTable, valueRange, parserMode, unitless, unitlessZero, std::forward<Args>(args)...);
-    }
-};
-
-
-// MARK: - Consumer definitions.
-
-// MARK: Integer
-
-template<typename IntType, IntegerValueRange integerRange>
-struct IntegerTypeRawConsumer {
-    using Result = std::optional<IntType>;
-
-    using FunctionToken = IntegerTypeRawKnownTokenTypeFunctionConsumer<IntType, integerRange>;
-    using NumberToken = IntegerTypeRawKnownTokenTypeNumberConsumer<IntType, integerRange>;
-};
-
-template<typename IntType, IntegerValueRange integerRange>
-struct IntegerTypeConsumer {
-    using Result = RefPtr<CSSPrimitiveValue>;
-
-    using FunctionToken = IntegerTypeKnownTokenTypeFunctionConsumer<IntType, integerRange>;
-    using NumberToken = IntegerTypeKnownTokenTypeNumberConsumer<IntType, integerRange>;
-};
-
-// MARK: Number
-
-template<typename T>
-struct NumberRawConsumer {
-    using Transformer = T;
-    using Result = typename Transformer::Result;
-
-    using FunctionToken = NumberRawKnownTokenTypeFunctionConsumer;
-    using NumberToken = NumberRawKnownTokenTypeNumberConsumer;
-};
-
-template<typename Transformer>
-struct NumberRawAllowingSymbolTableIdentConsumer : NumberRawConsumer<Transformer> {
-    using IdentToken = NumberRawKnownTokenTypeIdentConsumer;
-};
-
-struct NumberConsumer {
-    using Result = RefPtr<CSSPrimitiveValue>;
-
-    using FunctionToken = NumberCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer;
-    using NumberToken = NumberCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer;
-};
-
-// MARK: Percent
-
-template<typename T>
-struct PercentRawConsumer {
-    using Transformer = T;
-    using Result = typename Transformer::Result;
-
-    using FunctionToken = PercentRawKnownTokenTypeFunctionConsumer;
-    using PercentageToken = PercentRawKnownTokenTypePercentConsumer;
-};
-
-template<typename Transformer>
-struct PercentRawAllowingSymbolTableIdentConsumer : PercentRawConsumer<Transformer> {
-    using IdentToken = PercentRawKnownTokenTypeIdentConsumer;
-};
-
-struct PercentConsumer {
-    using Result = RefPtr<CSSPrimitiveValue>;
-
-    using FunctionToken = PercentCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer;
-    using PercentageToken = PercentCSSPrimitiveValueWithCalcWithKnownTokenTypePercentConsumer;
-};
-
-// MARK: Length
-
-template<typename T>
-struct LengthRawConsumer {
-    using Transformer = T;
-    using Result = typename Transformer::Result;
-
-    using FunctionToken = LengthRawKnownTokenTypeFunctionConsumer;
-    using DimensionToken = LengthRawKnownTokenTypeDimensionConsumer;
-    using NumberToken = LengthRawKnownTokenTypeNumberConsumer;
-};
-
-struct LengthConsumer {
-    using Result = RefPtr<CSSPrimitiveValue>;
-
-    using FunctionToken = LengthCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer;
-    using NumberToken = LengthCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer;
-    using DimensionToken = LengthCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer;
-};
-
-// MARK: Angle
-
-template<typename T>
-struct AngleRawConsumer {
-    using Transformer = T;
-    using Result = typename Transformer::Result;
-
-    using FunctionToken = AngleRawKnownTokenTypeFunctionConsumer;
-    using NumberToken = AngleRawKnownTokenTypeNumberConsumer;
-    using DimensionToken = AngleRawKnownTokenTypeDimensionConsumer;
-};
-
-struct AngleConsumer {
-    using Result = RefPtr<CSSPrimitiveValue>;
-
-    using FunctionToken = AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer;
-    using NumberToken = AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer;
-    using DimensionToken = AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer;
-};
-
-// MARK: Time
-
-struct TimeConsumer {
-    using Result = RefPtr<CSSPrimitiveValue>;
-
-    using FunctionToken = TimeCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer;
-    using NumberToken = TimeCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer;
-    using DimensionToken = TimeCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer;
-};
-
-// MARK: Resolution
-
-struct ResolutionConsumer {
-    using Result = RefPtr<CSSPrimitiveValue>;
-
-    using FunctionToken = ResolutionCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer;
-    using DimensionToken = ResolutionCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer;
-};
-
-// MARK: - Combination consumer definitions.
-
-// MARK: Angle + Percent
-
-struct AngleOrPercentConsumer {
-    using Result = RefPtr<CSSPrimitiveValue>;
-
-    using FunctionToken = SameTokenMetaConsumer<
-        IdentityTransformer<RefPtr<CSSPrimitiveValue>>,
-        AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer,
-        PercentCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer
-    >;
-    using NumberToken = AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer;
-    using PercentageToken = PercentCSSPrimitiveValueWithCalcWithKnownTokenTypePercentConsumer;
-    using DimensionToken = AngleCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer;
-};
-
-// MARK: Length + Percent
-
-template<typename T>
-struct LengthOrPercentRawConsumer {
-    using Transformer = T;
-    using Result = typename Transformer::Result;
-
-    using FunctionToken = SameTokenMetaConsumer<
-        Transformer,
-        PercentRawKnownTokenTypeFunctionConsumer,
-        LengthRawKnownTokenTypeFunctionConsumer
-    >;
-    using NumberToken = LengthRawKnownTokenTypeNumberConsumer;
-    using PercentageToken = PercentRawKnownTokenTypePercentConsumer;
-    using DimensionToken = LengthRawKnownTokenTypeDimensionConsumer;
-};
-
-// MARK: Angle + Number
-
-template<typename T>
-struct AngleOrNumberRawConsumer {
-    using Transformer = T;
-    using Result = typename Transformer::Result;
-
-    using FunctionToken = SameTokenMetaConsumer<
-        Transformer,
-        NumberRawKnownTokenTypeFunctionConsumer,
-        AngleRawKnownTokenTypeFunctionConsumer
-    >;
-    using NumberToken = NumberRawKnownTokenTypeNumberConsumer;
-    using DimensionToken = AngleRawKnownTokenTypeDimensionConsumer;
-};
-
-template<typename Transformer>
-struct AngleOrNumberRawAllowingSymbolTableIdentConsumer : AngleOrNumberRawConsumer<Transformer> {
-    using IdentToken = AngleOrNumberRawKnownTokenTypeIdentConsumer;
-};
-
-// MARK: Angle + Number + None
-
-template<typename Transformer>
-struct AngleOrNumberOrNoneRawConsumer : AngleOrNumberRawConsumer<Transformer> {
-    using IdentToken = NoneRawKnownTokenTypeIdentConsumer;
-};
-
-template<typename Transformer>
-struct AngleOrNumberOrNoneRawAllowingSymbolTableIdentConsumer : AngleOrNumberRawConsumer<Transformer> {
-    using IdentToken = SameTokenMetaConsumer<
-        Transformer,
-        NoneRawKnownTokenTypeIdentConsumer,
-        AngleOrNumberRawKnownTokenTypeIdentConsumer
-    >;
-};
-
-// MARK: Number + Percent
-
-template<typename T>
-struct NumberOrPercentRawConsumer {
-    using Transformer = T;
-    using Result = typename Transformer::Result;
-
-    using FunctionToken = SameTokenMetaConsumer<
-        Transformer,
-        PercentRawKnownTokenTypeFunctionConsumer,
-        NumberRawKnownTokenTypeFunctionConsumer
-    >;
-    using NumberToken = NumberRawKnownTokenTypeNumberConsumer;
-    using PercentageToken = PercentRawKnownTokenTypePercentConsumer;
-};
-
-
-template<typename Transformer>
-struct NumberOrPercentRawAllowingSymbolTableIdentConsumer : NumberOrPercentRawConsumer<Transformer> {
-    using IdentToken = NumberOrPercentRawKnownTokenTypeIdentConsumer;
-};
-
-// MARK: Number + None
-
-template<typename Transformer>
-struct NumberOrNoneRawConsumer : NumberRawConsumer<Transformer> {
-    using IdentToken = NoneRawKnownTokenTypeIdentConsumer;
-};
-
-template<typename Transformer>
-struct NumberOrNoneRawAllowingSymbolTableIdentConsumer : NumberRawConsumer<Transformer> {
-    using IdentToken = SameTokenMetaConsumer<
-        Transformer,
-        NoneRawKnownTokenTypeIdentConsumer,
-        NumberRawKnownTokenTypeIdentConsumer
-    >;
-};
-
-// MARK: Percent + None
-
-template<typename Transformer>
-struct PercentOrNoneRawConsumer : PercentRawConsumer<Transformer> {
-    using IdentToken = NoneRawKnownTokenTypeIdentConsumer;
-};
-
-template<typename Transformer>
-struct PercentOrNoneRawAllowingSymbolTableIdentConsumer : PercentRawConsumer<Transformer> {
-    using IdentToken = SameTokenMetaConsumer<
-        Transformer,
-        NoneRawKnownTokenTypeIdentConsumer,
-        PercentRawKnownTokenTypeIdentConsumer
-    >;
-};
-
-// MARK: Number + Percent + None
-
-template<typename Transformer>
-struct NumberOrPercentOrNoneRawConsumer : NumberOrPercentRawConsumer<Transformer> {
-    using IdentToken = NoneRawKnownTokenTypeIdentConsumer;
-};
-
-template<typename Transformer>
-struct NumberOrPercentOrNoneRawAllowingSymbolTableIdentConsumer : NumberOrPercentRawConsumer<Transformer> {
-    using IdentToken = SameTokenMetaConsumer<
-        Transformer,
-        NoneRawKnownTokenTypeIdentConsumer,
-        NumberOrPercentRawKnownTokenTypeIdentConsumer
-    >;
-};
-
 // MARK: Image Set Resolution + Type
 
 struct ImageSetResolutionOrTypeConsumer {
@@ -1412,225 +154,6 @@ struct ImageSetResolutionOrTypeConsumer {
 
     using DimensionToken = ResolutionCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer;
 };
-
-
-// MARK: - Consumer functions - utilize consumer definitions above, giving more targetted interfaces and allowing exposure to other files.
-
-template<typename IntType, IntegerValueRange integerRange> std::optional<IntType> consumeIntegerTypeRaw(CSSParserTokenRange& range)
-{
-    return consumeMetaConsumer<IntegerTypeRawConsumer<IntType, integerRange>>(range, { }, ValueRange::All, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-template<typename IntType, IntegerValueRange integerRange> RefPtr<CSSPrimitiveValue> consumeIntegerType(CSSParserTokenRange& range)
-{
-    return consumeMetaConsumer<IntegerTypeConsumer<IntType, integerRange>>(range, { }, ValueRange::All, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-std::optional<int> consumeIntegerRaw(CSSParserTokenRange& range)
-{
-    return consumeIntegerTypeRaw<int, IntegerValueRange::All>(range);
-}
-
-RefPtr<CSSPrimitiveValue> consumeInteger(CSSParserTokenRange& range)
-{
-    return consumeIntegerType<int, IntegerValueRange::All>(range);
-}
-
-std::optional<int> consumeNonNegativeIntegerRaw(CSSParserTokenRange& range)
-{
-    return consumeIntegerTypeRaw<int, IntegerValueRange::NonNegative>(range);
-}
-
-RefPtr<CSSPrimitiveValue> consumeNonNegativeInteger(CSSParserTokenRange& range)
-{
-    return consumeIntegerType<int, IntegerValueRange::NonNegative>(range);
-}
-
-std::optional<unsigned> consumePositiveIntegerRaw(CSSParserTokenRange& range)
-{
-    return consumeIntegerTypeRaw<unsigned, IntegerValueRange::Positive>(range);
-}
-
-RefPtr<CSSPrimitiveValue> consumePositiveInteger(CSSParserTokenRange& range)
-{
-    return consumeIntegerType<unsigned, IntegerValueRange::Positive>(range);
-}
-
-RefPtr<CSSPrimitiveValue> consumeInteger(CSSParserTokenRange& range, IntegerValueRange valueRange)
-{
-    switch (valueRange) {
-    case IntegerValueRange::All:
-        return consumeInteger(range);
-    case IntegerValueRange::Positive:
-        return consumePositiveInteger(range);
-    case IntegerValueRange::NonNegative:
-        return consumeNonNegativeInteger(range);
-    }
-    RELEASE_ASSERT_NOT_REACHED();
-}
-
-std::optional<NumberRaw> consumeNumberRaw(CSSParserTokenRange& range, ValueRange valueRange)
-{
-    return consumeMetaConsumer<NumberRawConsumer<RawIdentityTransformer<NumberRaw>>>(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-RefPtr<CSSPrimitiveValue> consumeNumber(CSSParserTokenRange& range, ValueRange valueRange)
-{
-    return consumeMetaConsumer<NumberConsumer>(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-static std::optional<PercentRaw> consumePercentRaw(CSSParserTokenRange& range)
-{
-    return consumeMetaConsumer<PercentRawConsumer<RawIdentityTransformer<PercentRaw>>>(range, { }, ValueRange::All, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-RefPtr<CSSPrimitiveValue> consumePercent(CSSParserTokenRange& range, ValueRange valueRange)
-{
-    return consumeMetaConsumer<PercentConsumer>(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-RefPtr<CSSPrimitiveValue> consumeLength(CSSParserTokenRange& range, CSSParserMode parserMode, ValueRange valueRange, UnitlessQuirk unitless)
-{
-    return consumeMetaConsumer<LengthConsumer>(range, { }, valueRange, parserMode, unitless, UnitlessZeroQuirk::Allow);
-}
-
-#if ENABLE(VARIATION_FONTS)
-static std::optional<AngleRaw> consumeAngleRaw(CSSParserTokenRange& range, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-{
-    return consumeMetaConsumer<AngleRawConsumer<RawIdentityTransformer<AngleRaw>>>(range, { }, ValueRange::All, parserMode, unitless, unitlessZero);
-}
-#endif
-
-RefPtr<CSSPrimitiveValue> consumeAngle(CSSParserTokenRange& range, CSSParserMode parserMode, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero)
-{
-    return consumeMetaConsumer<AngleConsumer>(range, { }, ValueRange::All, parserMode, unitless, unitlessZero);
-}
-
-RefPtr<CSSPrimitiveValue> consumeTime(CSSParserTokenRange& range, CSSParserMode parserMode, ValueRange valueRange, UnitlessQuirk unitless)
-{
-    return consumeMetaConsumer<TimeConsumer>(range, { }, valueRange, parserMode, unitless, UnitlessZeroQuirk::Forbid);
-}
-
-RefPtr<CSSPrimitiveValue> consumeResolution(CSSParserTokenRange& range)
-{
-    return consumeMetaConsumer<ResolutionConsumer>(range, { }, ValueRange::NonNegative, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-static RefPtr<CSSPrimitiveValue> consumeAngleOrPercent(CSSParserTokenRange& range, CSSParserMode parserMode)
-{
-    return consumeMetaConsumer<AngleOrPercentConsumer>(range, { }, ValueRange::All, parserMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Allow);
-}
-
-static std::optional<LengthOrPercentRaw> consumeLengthOrPercentRaw(CSSParserTokenRange& range, CSSParserMode parserMode)
-{
-    return consumeMetaConsumer<LengthOrPercentRawConsumer<RawIdentityTransformer<LengthOrPercentRaw>>>(range, { }, ValueRange::NonNegative, parserMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-// FIXME: This doesn't work with the current scheme due to the NegativePercentagePolicy parameter
-RefPtr<CSSPrimitiveValue> consumeLengthOrPercent(CSSParserTokenRange& range, CSSParserMode parserMode, ValueRange valueRange, UnitlessQuirk unitless, UnitlessZeroQuirk unitlessZero, NegativePercentagePolicy negativePercentagePolicy)
-{
-    auto& token = range.peek();
-
-    switch (token.type()) {
-    case FunctionToken: {
-        // FIXME: Should this be using trying to generate the calc with both Length and Percent destination category types?
-        CalcParser parser(range, CalculationCategory::Length, valueRange, { }, negativePercentagePolicy);
-        if (auto calculation = parser.value(); calculation && canConsumeCalcValue(calculation->category(), parserMode))
-            return parser.consumeValue();
-        break;
-    }
-
-    case DimensionToken:
-        return LengthCSSPrimitiveValueWithCalcWithKnownTokenTypeDimensionConsumer::consume(range, { }, valueRange, parserMode, unitless, unitlessZero);
-
-    case NumberToken:
-        return LengthCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer::consume(range, { }, valueRange, parserMode, unitless, unitlessZero);
-
-    case PercentageToken:
-        return PercentCSSPrimitiveValueWithCalcWithKnownTokenTypePercentConsumer::consume(range, { }, valueRange, parserMode, unitless, unitlessZero);
-
-    default:
-        break;
-    }
-
-    return nullptr;
-}
-
-template<typename Transformer = RawIdentityTransformer<AngleOrNumberOrNoneRaw>>
-static auto consumeAngleOrNumberOrNoneRaw(CSSParserTokenRange& range, CSSParserMode parserMode) -> typename Transformer::Result
-{
-    return consumeMetaConsumer<AngleOrNumberOrNoneRawConsumer<Transformer>>(range, { }, ValueRange::All, parserMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-template<typename Transformer = RawIdentityTransformer<AngleOrNumberOrNoneRaw>>
-static auto consumeAngleOrNumberOrNoneRawAllowingSymbolTableIdent(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, CSSParserMode parserMode) -> typename Transformer::Result
-{
-    return consumeMetaConsumer<AngleOrNumberOrNoneRawAllowingSymbolTableIdentConsumer<Transformer>>(range, symbolTable, ValueRange::All, parserMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-template<typename Transformer = RawIdentityTransformer<NumberOrPercentRaw>>
-static auto consumeNumberOrPercentRaw(CSSParserTokenRange& range, ValueRange valueRange = ValueRange::All) -> typename Transformer::Result
-{
-    return consumeMetaConsumer<NumberOrPercentRawConsumer<Transformer>>(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-template<typename Transformer = RawIdentityTransformer<NumberOrNoneRaw>>
-static auto consumeNumberOrNoneRaw(CSSParserTokenRange& range, ValueRange valueRange = ValueRange::All) -> typename Transformer::Result
-{
-    return consumeMetaConsumer<NumberOrNoneRawConsumer<Transformer>>(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-template<typename Transformer = RawIdentityTransformer<PercentOrNoneRaw>>
-static auto consumePercentOrNoneRaw(CSSParserTokenRange& range, ValueRange valueRange = ValueRange::All) -> typename Transformer::Result
-{
-    return consumeMetaConsumer<PercentOrNoneRawConsumer<Transformer>>(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-template<typename Transformer = RawIdentityTransformer<PercentOrNoneRaw>>
-static auto consumePercentOrNoneRawAllowingSymbolTableIdent(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange = ValueRange::All) -> typename Transformer::Result
-{
-    return consumeMetaConsumer<PercentOrNoneRawAllowingSymbolTableIdentConsumer<Transformer>>(range, symbolTable, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-template<typename Transformer = RawIdentityTransformer<NumberOrPercentOrNoneRaw>>
-static auto consumeNumberOrPercentOrNoneRaw(CSSParserTokenRange& range, ValueRange valueRange = ValueRange::All) -> typename Transformer::Result
-{
-    return consumeMetaConsumer<NumberOrPercentOrNoneRawConsumer<Transformer>>(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-template<typename Transformer = RawIdentityTransformer<NumberOrPercentOrNoneRaw>>
-static auto consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable, ValueRange valueRange = ValueRange::All) -> typename Transformer::Result
-{
-    return consumeMetaConsumer<NumberOrPercentOrNoneRawAllowingSymbolTableIdentConsumer<Transformer>>(range, symbolTable, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-}
-
-// FIXME: This needs a more clear name to indicate its behavior of dividing percents by 100 only if an explicit percent token, not if a result of a calc().
-RefPtr<CSSPrimitiveValue> consumeNumberOrPercent(CSSParserTokenRange& range, ValueRange valueRange)
-{
-    auto& token = range.peek();
-
-    switch (token.type()) {
-    case FunctionToken:
-        if (RefPtr value = NumberCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer::consume(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid))
-            return value;
-        return PercentCSSPrimitiveValueWithCalcWithKnownTokenTypeFunctionConsumer::consume(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-
-    case NumberToken:
-        return NumberCSSPrimitiveValueWithCalcWithKnownTokenTypeNumberConsumer::consume(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid);
-
-    case PercentageToken:
-        if (auto percentRaw = PercentRawKnownTokenTypePercentConsumer::consume(range, { }, valueRange, CSSParserMode::HTMLStandardMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Forbid))
-            return CSSPrimitiveValue::create(percentRaw->value / 100.0);
-        break;
-
-    default:
-        break;
-    }
-
-    return nullptr;
-}
-
-// MARK: - Non-primitive consumers.
 
 static std::optional<double> consumeFontWeightNumberRaw(CSSParserTokenRange& range)
 {
@@ -1677,35 +200,6 @@ RefPtr<CSSPrimitiveValue> consumeFontWeightNumber(CSSParserTokenRange& range)
     if (auto result = consumeFontWeightNumberRaw(range))
         return CSSPrimitiveValue::create(*result);
     return nullptr;
-}
-
-std::optional<CSSValueID> consumeIdentRaw(CSSParserTokenRange& range)
-{
-    if (range.peek().type() != IdentToken)
-        return std::nullopt;
-    return range.consumeIncludingWhitespace().id();
-}
-
-RefPtr<CSSPrimitiveValue> consumeIdent(CSSParserTokenRange& range)
-{
-    if (auto result = consumeIdentRaw(range))
-        return CSSPrimitiveValue::create(*result);
-    return nullptr;
-}
-
-static std::optional<CSSValueID> consumeIdentRangeRaw(CSSParserTokenRange& range, CSSValueID lower, CSSValueID upper)
-{
-    if (range.peek().id() < lower || range.peek().id() > upper)
-        return std::nullopt;
-    return consumeIdentRaw(range);
-}
-
-RefPtr<CSSPrimitiveValue> consumeIdentRange(CSSParserTokenRange& range, CSSValueID lower, CSSValueID upper)
-{
-    auto value = consumeIdentRangeRaw(range, lower, upper);
-    if (!value)
-        return nullptr;
-    return CSSPrimitiveValue::create(*value);
 }
 
 static String consumeCustomIdentRaw(CSSParserTokenRange& range, bool shouldLowercase = false)
@@ -1767,1597 +261,6 @@ RefPtr<CSSPrimitiveValue> consumeURL(CSSParserTokenRange& range)
     if (url.isNull())
         return nullptr;
     return CSSPrimitiveValue::createURI(url.toString());
-}
-
-static Color consumeOriginColorRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    auto value = consumeColor(args, context);
-    if (!value)
-        return { };
-
-    if (value->isColor())
-        return value->color();
-
-    // FIXME: We don't know how to deal with unresolved origin color at parse time.
-    // https://bugs.webkit.org/show_bug.cgi?id=245970
-    if (value->isUnresolvedColor())
-        return { };
-
-    ASSERT(value->isValueID());
-    auto keyword = value->valueID();
-
-    // FIXME: We don't have enough context in the parser to resolving a system keyword
-    // correctly. We should package up the relative color parameters and resolve the
-    // whole thing at the appropriate time when the origin color is a system keyword.
-    if (StyleColor::isSystemColorKeyword(keyword))
-        return { };
-
-    return StyleColor::colorFromKeyword(keyword, { });
-}
-
-static std::optional<double> consumeOptionalAlphaRaw(CSSParserTokenRange& range)
-{
-    if (!consumeSlashIncludingWhitespace(range))
-        return 1.0;
-
-    if (auto alphaParameter = consumeNumberOrPercentOrNoneRaw(range)) {
-        return WTF::switchOn(*alphaParameter,
-            [] (NumberRaw number) { return std::clamp(number.value, 0.0, 1.0); },
-            [] (PercentRaw percent) { return std::clamp(percent.value / 100.0, 0.0, 1.0); },
-            [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-        );
-    }
-
-    return std::nullopt;
-}
-
-static std::optional<double> consumeOptionalAlphaRawAllowingSymbolTableIdent(CSSParserTokenRange& range, const CSSCalcSymbolTable& symbolTable)
-{
-    if (!consumeSlashIncludingWhitespace(range))
-        return 1.0;
-
-    if (auto alphaParameter = consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(range, symbolTable, ValueRange::All)) {
-        return WTF::switchOn(*alphaParameter,
-            [] (NumberRaw number) { return std::clamp(number.value, 0.0, 1.0); },
-            [] (PercentRaw percent) { return std::clamp(percent.value / 100.0, 0.0, 1.0); },
-            [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-        );
-    }
-
-    return std::nullopt;
-}
-
-static uint8_t normalizeRGBComponentToSRGBAByte(NumberRaw value)
-{
-    return convertPrescaledSRGBAFloatToSRGBAByte(value.value);
-}
-
-static uint8_t normalizeRGBComponentToSRGBAByte(PercentRaw value)
-{
-    return convertPrescaledSRGBAFloatToSRGBAByte(value.value / 100.0 * 255.0);
-}
-
-enum class RGBOrHSLSeparatorSyntax { Commas, WhitespaceSlash };
-
-static bool consumeRGBOrHSLSeparator(CSSParserTokenRange& args, RGBOrHSLSeparatorSyntax syntax)
-{
-    if (syntax == RGBOrHSLSeparatorSyntax::Commas)
-        return consumeCommaIncludingWhitespace(args);
-    return true;
-}
-
-static bool consumeRGBOrHSLAlphaSeparator(CSSParserTokenRange& args, RGBOrHSLSeparatorSyntax syntax)
-{
-    if (syntax == RGBOrHSLSeparatorSyntax::Commas)
-        return consumeCommaIncludingWhitespace(args);
-    return consumeSlashIncludingWhitespace(args);
-}
-
-static std::optional<double> consumeRGBOrHSLOptionalAlpha(CSSParserTokenRange& args, RGBOrHSLSeparatorSyntax syntax)
-{
-    if (!consumeRGBOrHSLAlphaSeparator(args, syntax))
-        return 1.0;
-
-    if (auto alphaParameter = consumeNumberOrPercentOrNoneRaw(args)) {
-        return WTF::switchOn(*alphaParameter,
-            [] (NumberRaw number) { return std::clamp(number.value, 0.0, 1.0); },
-            [] (PercentRaw percent) { return std::clamp(percent.value / 100.0, 0.0, 1.0); },
-            [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-        );
-    }
-
-    return std::nullopt;
-}
-
-static Color parseRelativeRGBParametersRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    ASSERT(args.peek().id() == CSSValueFrom);
-    consumeIdentRaw(args);
-
-    auto originColor = consumeOriginColorRaw(args, context);
-    if (!originColor.isValid())
-        return { };
-
-    auto originColorAsSRGB = originColor.toColorTypeLossy<SRGBA<float>>().resolved();
-
-    CSSCalcSymbolTable symbolTable {
-        { CSSValueR, CSSUnitType::CSS_PERCENTAGE, originColorAsSRGB.red * 100.0 },
-        { CSSValueG, CSSUnitType::CSS_PERCENTAGE, originColorAsSRGB.green * 100.0 },
-        { CSSValueB, CSSUnitType::CSS_PERCENTAGE, originColorAsSRGB.blue * 100.0 },
-        { CSSValueAlpha, CSSUnitType::CSS_PERCENTAGE, originColorAsSRGB.alpha * 100.0 }
-    };
-
-    auto red = consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable);
-    if (!red)
-        return { };
-
-    auto green = consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable);
-    if (!green)
-        return { };
-
-    auto blue = consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable);
-    if (!blue)
-        return { };
-
-    auto alpha = consumeOptionalAlphaRawAllowingSymbolTableIdent(args, symbolTable);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    if (std::holds_alternative<NoneRaw>(*red) || std::holds_alternative<NoneRaw>(*green) || std::holds_alternative<NoneRaw>(*blue) || std::isnan(*alpha)) {
-        auto normalizeComponentAllowingNone = [] (auto component) {
-            return WTF::switchOn(component,
-                [] (PercentRaw percent) { return std::clamp(percent.value / 100.0, 0.0, 1.0); },
-                [] (NumberRaw number) { return std::clamp(number.value / 255.0, 0.0, 1.0); },
-                [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-            );
-        };
-
-        auto normalizedRed = normalizeComponentAllowingNone(*red);
-        auto normalizedGreen = normalizeComponentAllowingNone(*green);
-        auto normalizedBlue = normalizeComponentAllowingNone(*blue);
-
-        // If any component uses "none", we store the value as a SRGBA<float> to allow for storage of the special value as NaN.
-        return SRGBA<float> { static_cast<float>(normalizedRed), static_cast<float>(normalizedGreen), static_cast<float>(normalizedBlue), static_cast<float>(*alpha) };
-    }
-
-    auto normalizeComponentDisallowingNone = [] (auto component) {
-        return WTF::switchOn(component,
-            [] (NumberRaw number) -> uint8_t { return normalizeRGBComponentToSRGBAByte(number); },
-            [] (PercentRaw percent) -> uint8_t { return normalizeRGBComponentToSRGBAByte(percent); },
-            [] (NoneRaw) -> uint8_t { ASSERT_NOT_REACHED(); return 0; }
-        );
-    };
-
-    auto normalizedRed = normalizeComponentDisallowingNone(*red);
-    auto normalizedGreen = normalizeComponentDisallowingNone(*green);
-    auto normalizedBlue = normalizeComponentDisallowingNone(*blue);
-    auto normalizedAlpha = convertFloatAlphaTo<uint8_t>(*alpha);
-
-    return SRGBA<uint8_t> { normalizedRed, normalizedGreen, normalizedBlue, normalizedAlpha };
-}
-
-static Color parseNonRelativeRGBParametersRaw(CSSParserTokenRange& args)
-{
-    struct Component {
-        enum class Type { Number, Percentage, Unknown };
-
-        double value;
-        Type type;
-    };
-
-    auto consumeComponent = [](auto& args, auto previousComponentType) -> std::optional<Component> {
-        switch (previousComponentType) {
-        case Component::Type::Number:
-            if (auto component = consumeNumberOrNoneRaw(args)) {
-                return WTF::switchOn(*component,
-                    [] (NumberRaw number) -> Component { return { number.value, Component::Type::Number }; },
-                    [] (NoneRaw) -> Component { return { std::numeric_limits<double>::quiet_NaN(), Component::Type::Number }; }
-                );
-            }
-            return std::nullopt;
-        case Component::Type::Percentage:
-            if (auto component = consumePercentOrNoneRaw(args)) {
-                return WTF::switchOn(*component,
-                    [] (PercentRaw percent) -> Component  { return { percent.value, Component::Type::Percentage }; },
-                    [] (NoneRaw) -> Component { return { std::numeric_limits<double>::quiet_NaN(), Component::Type::Percentage }; }
-                );
-            }
-            return std::nullopt;
-        case Component::Type::Unknown:
-            if (auto component = consumeNumberOrPercentOrNoneRaw(args)) {
-                return WTF::switchOn(*component,
-                    [] (NumberRaw number) -> Component { return { number.value, Component::Type::Number }; },
-                    [] (PercentRaw percent) -> Component  { return { percent.value, Component::Type::Percentage }; },
-                    [] (NoneRaw) -> Component { return { std::numeric_limits<double>::quiet_NaN(), Component::Type::Unknown }; }
-                );
-            }
-            return std::nullopt;
-        }
-        RELEASE_ASSERT_NOT_REACHED();
-    };
-
-    auto red = consumeComponent(args, Component::Type::Unknown);
-    if (!red)
-        return { };
-
-    auto syntax = consumeCommaIncludingWhitespace(args) ? RGBOrHSLSeparatorSyntax::Commas : RGBOrHSLSeparatorSyntax::WhitespaceSlash;
-
-    auto green = consumeComponent(args, red->type);
-    if (!green)
-        return { };
-
-    if (!consumeRGBOrHSLSeparator(args, syntax))
-        return { };
-
-    auto blue = consumeComponent(args, green->type);
-    if (!blue)
-        return { };
-
-    auto resolvedComponentType = blue->type;
-
-    auto alpha = consumeRGBOrHSLOptionalAlpha(args, syntax);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    if (std::isnan(red->value) || std::isnan(green->value) || std::isnan(blue->value) || std::isnan(*alpha)) {
-        // "none" values are only allowed with the WhitespaceSlash syntax.
-        if (syntax != RGBOrHSLSeparatorSyntax::WhitespaceSlash)
-            return { };
-
-        auto normalizeNumber = [] (double number) { return std::isnan(number) ? number : std::clamp(number / 255.0, 0.0, 1.0); };
-        auto normalizePercent = [] (double percent) { return std::isnan(percent) ? percent : std::clamp(percent / 100.0, 0.0, 1.0); };
-
-        // If any component uses "none", we store the value as a SRGBA<float> to allow for storage of the special value as NaN.
-        switch (resolvedComponentType) {
-        case Component::Type::Number:
-            return SRGBA<float> { static_cast<float>(normalizeNumber(red->value)), static_cast<float>(normalizeNumber(green->value)), static_cast<float>(normalizeNumber(blue->value)), static_cast<float>(*alpha) };
-        case Component::Type::Percentage:
-            return SRGBA<float> { static_cast<float>(normalizePercent(red->value)), static_cast<float>(normalizePercent(green->value)), static_cast<float>(normalizePercent(blue->value)), static_cast<float>(*alpha) };
-        case Component::Type::Unknown:
-            return SRGBA<float> { static_cast<float>(red->value), static_cast<float>(green->value), static_cast<float>(blue->value), static_cast<float>(*alpha) };
-        }
-
-        ASSERT_NOT_REACHED();
-        return { };
-    }
-
-    switch (resolvedComponentType) {
-    case Component::Type::Number:
-        return SRGBA<uint8_t> { normalizeRGBComponentToSRGBAByte(NumberRaw { red->value }), normalizeRGBComponentToSRGBAByte(NumberRaw { green->value }), normalizeRGBComponentToSRGBAByte(NumberRaw { blue->value }), convertFloatAlphaTo<uint8_t>(*alpha) };
-    case Component::Type::Percentage:
-        return SRGBA<uint8_t> { normalizeRGBComponentToSRGBAByte(PercentRaw { red->value }), normalizeRGBComponentToSRGBAByte(PercentRaw { green->value }), normalizeRGBComponentToSRGBAByte(PercentRaw { blue->value }), convertFloatAlphaTo<uint8_t>(*alpha) };
-    case Component::Type::Unknown:
-        // The only way the resolvedComponentType can be Component::Type::Unknown is if all the components are "none", which is handled above.
-        ASSERT_NOT_REACHED();
-        return { };
-    }
-
-    ASSERT_NOT_REACHED();
-    return { };
-}
-
-enum class RGBFunctionMode { RGB, RGBA };
-
-template<RGBFunctionMode Mode> static Color parseRGBParametersRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    ASSERT(range.peek().functionId() == (Mode == RGBFunctionMode::RGB ? CSSValueRgb : CSSValueRgba));
-    auto args = consumeFunction(range);
-
-    if constexpr (Mode == RGBFunctionMode::RGB) {
-        if (args.peek().id() == CSSValueFrom)
-            return parseRelativeRGBParametersRaw(args, context);
-    }
-    return parseNonRelativeRGBParametersRaw(args);
-}
-
-static Color colorByNormalizingHSLComponents(AngleOrNumberOrNoneRaw hue, PercentOrNoneRaw saturation, PercentOrNoneRaw lightness, double alpha, RGBOrHSLSeparatorSyntax syntax)
-{
-    auto normalizedHue = WTF::switchOn(hue,
-        [] (AngleRaw angle) { return CSSPrimitiveValue::computeDegrees(angle.type, angle.value); },
-        [] (NumberRaw number) { return number.value; },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-    auto normalizedSaturation = WTF::switchOn(saturation,
-        [] (PercentRaw percent) { return std::clamp(percent.value, 0.0, 100.0); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-    auto normalizedLightness = WTF::switchOn(lightness,
-        [] (PercentRaw percent) { return std::clamp(percent.value, 0.0, 100.0); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-
-    if (std::isnan(normalizedHue) || std::isnan(normalizedSaturation) || std::isnan(normalizedLightness) || std::isnan(alpha)) {
-        // "none" values are only allowed with the WhitespaceSlash syntax.
-        if (syntax != RGBOrHSLSeparatorSyntax::WhitespaceSlash)
-            return { };
-
-        // If any component uses "none", we store the value as a HSLA<float> to allow for storage of the special value as NaN.
-        return HSLA<float> { static_cast<float>(normalizedHue), static_cast<float>(normalizedSaturation), static_cast<float>(normalizedLightness), static_cast<float>(alpha) };
-    }
-
-    if (normalizedHue < 0.0 || normalizedHue > 360.0) {
-        // If hue is not in the [0, 360] range, we store the value as a HSLA<float> to allow for correct interpolation
-        // using the "specified" hue interpolation method.
-        return HSLA<float> { static_cast<float>(normalizedHue), static_cast<float>(normalizedSaturation), static_cast<float>(normalizedLightness), static_cast<float>(alpha) };
-    }
-
-    // NOTE: The explicit conversion to SRGBA<uint8_t> is an intentional performance optimization that allows storing the
-    // color with no extra allocation for an extended color object. This is permissible due to the historical requirement
-    // that HSLA colors serialize using the legacy color syntax (rgb()/rgba()) and historically have used the 8-bit rgba
-    // internal representation in engines.
-    return convertColor<SRGBA<uint8_t>>(HSLA<float> { static_cast<float>(normalizedHue), static_cast<float>(normalizedSaturation), static_cast<float>(normalizedLightness), static_cast<float>(alpha) });
-}
-
-static Color parseRelativeHSLParametersRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    ASSERT(args.peek().id() == CSSValueFrom);
-    consumeIdentRaw(args);
-
-    auto originColor = consumeOriginColorRaw(args, context);
-    if (!originColor.isValid())
-        return { };
-
-    auto originColorAsHSL = originColor.toColorTypeLossy<HSLA<float>>().resolved();
-
-    CSSCalcSymbolTable symbolTable {
-        { CSSValueH, CSSUnitType::CSS_DEG, originColorAsHSL.hue },
-        { CSSValueS, CSSUnitType::CSS_PERCENTAGE, originColorAsHSL.saturation },
-        { CSSValueL, CSSUnitType::CSS_PERCENTAGE, originColorAsHSL.lightness },
-        { CSSValueAlpha, CSSUnitType::CSS_PERCENTAGE, originColorAsHSL.alpha * 100.0 }
-    };
-
-    auto hue = consumeAngleOrNumberOrNoneRawAllowingSymbolTableIdent(args, symbolTable, context.mode);
-    if (!hue)
-        return { };
-
-    auto saturation = consumePercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable);
-    if (!saturation)
-        return { };
-
-    auto lightness = consumePercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable);
-    if (!lightness)
-        return { };
-
-    auto alpha = consumeOptionalAlphaRawAllowingSymbolTableIdent(args, symbolTable);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    return colorByNormalizingHSLComponents(*hue, *saturation, *lightness, *alpha, RGBOrHSLSeparatorSyntax::WhitespaceSlash);
-}
-
-static Color parseNonRelativeHSLParametersRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    auto hue = consumeAngleOrNumberOrNoneRaw(args, context.mode);
-    if (!hue)
-        return { };
-
-    auto syntax = consumeCommaIncludingWhitespace(args) ? RGBOrHSLSeparatorSyntax::Commas : RGBOrHSLSeparatorSyntax::WhitespaceSlash;
-
-    auto saturation = consumePercentOrNoneRaw(args);
-    if (!saturation)
-        return { };
-
-    if (!consumeRGBOrHSLSeparator(args, syntax))
-        return { };
-
-    auto lightness = consumePercentOrNoneRaw(args);
-    if (!lightness)
-        return { };
-
-    auto alpha = consumeRGBOrHSLOptionalAlpha(args, syntax);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    return colorByNormalizingHSLComponents(*hue, *saturation, *lightness, *alpha, syntax);
-}
-
-enum class HSLFunctionMode { HSL, HSLA };
-
-template<HSLFunctionMode Mode> static Color parseHSLParametersRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    ASSERT(range.peek().functionId() == (Mode == HSLFunctionMode::HSL ? CSSValueHsl : CSSValueHsla));
-    auto args = consumeFunction(range);
-
-    if constexpr (Mode == HSLFunctionMode::HSL) {
-        if (args.peek().id() == CSSValueFrom)
-            return parseRelativeHSLParametersRaw(args, context);
-    }
-
-    return parseNonRelativeHSLParametersRaw(args, context);
-}
-
-template<typename ConsumerForHue, typename ConsumerForWhitenessAndBlackness, typename ConsumerForAlpha>
-static Color parseHWBParametersRaw(CSSParserTokenRange& args, ConsumerForHue&& hueConsumer, ConsumerForWhitenessAndBlackness&& whitenessAndBlacknessConsumer, ConsumerForAlpha&& alphaConsumer)
-{
-    auto hue = hueConsumer(args);
-    if (!hue)
-        return { };
-
-    auto whiteness = whitenessAndBlacknessConsumer(args);
-    if (!whiteness)
-        return { };
-
-    auto blackness = whitenessAndBlacknessConsumer(args);
-    if (!blackness)
-        return { };
-
-    auto alpha = alphaConsumer(args);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    auto normalizedHue = WTF::switchOn(*hue,
-        [] (AngleRaw angle) { return CSSPrimitiveValue::computeDegrees(angle.type, angle.value); },
-        [] (NumberRaw number) { return number.value; },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-    auto clampedWhiteness = WTF::switchOn(*whiteness,
-        [] (PercentRaw percent) { return std::clamp(percent.value, 0.0, 100.0); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-    auto clampedBlackness = WTF::switchOn(*blackness,
-        [] (PercentRaw percent) { return std::clamp(percent.value, 0.0, 100.0); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-
-    if (std::isnan(normalizedHue) || std::isnan(clampedWhiteness) || std::isnan(clampedBlackness) || std::isnan(*alpha)) {
-        auto [normalizedWhitness, normalizedBlackness] = normalizeClampedWhitenessBlacknessAllowingNone(clampedWhiteness, clampedBlackness);
-
-        // If any component uses "none", we store the value as a HWBA<float> to allow for storage of the special value as NaN.
-        return HWBA<float> { static_cast<float>(normalizedHue), static_cast<float>(normalizedWhitness), static_cast<float>(normalizedBlackness), static_cast<float>(*alpha) };
-    }
-
-    auto [normalizedWhitness, normalizedBlackness] = normalizeClampedWhitenessBlacknessDisallowingNone(clampedWhiteness, clampedBlackness);
-
-    if (normalizedHue < 0.0 || normalizedHue > 360.0) {
-        // If 'hue' is not in the [0, 360] range, we store the value as a HWBA<float> to allow for correct interpolation
-        // using the "specified" hue interpolation method.
-        return HWBA<float> { static_cast<float>(normalizedHue), static_cast<float>(normalizedWhitness), static_cast<float>(normalizedBlackness), static_cast<float>(*alpha) };
-    }
-
-    // NOTE: The explicit conversion to SRGBA<uint8_t> is an intentional performance optimization that allows storing the
-    // color with no extra allocation for an extended color object. This is permissible due to the historical requirement
-    // that HWBA colors serialize using the legacy color syntax (rgb()/rgba()) and historically have used the 8-bit rgba
-    // internal representation in engines.
-    return convertColor<SRGBA<uint8_t>>(HWBA<float> { static_cast<float>(normalizedHue), static_cast<float>(normalizedWhitness), static_cast<float>(normalizedBlackness), static_cast<float>(*alpha) });
-}
-
-static Color parseRelativeHWBParametersRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    ASSERT(args.peek().id() == CSSValueFrom);
-    consumeIdentRaw(args);
-
-    auto originColor = consumeOriginColorRaw(args, context);
-    if (!originColor.isValid())
-        return { };
-
-    auto originColorAsHWB = originColor.toColorTypeLossy<HWBA<float>>().resolved();
-
-    CSSCalcSymbolTable symbolTable {
-        { CSSValueH, CSSUnitType::CSS_DEG, originColorAsHWB.hue },
-        { CSSValueW, CSSUnitType::CSS_PERCENTAGE, originColorAsHWB.whiteness },
-        { CSSValueB, CSSUnitType::CSS_PERCENTAGE, originColorAsHWB.blackness },
-        { CSSValueAlpha, CSSUnitType::CSS_PERCENTAGE, originColorAsHWB.alpha * 100.0 }
-    };
-
-    auto hueConsumer = [&symbolTable, &context](auto& args) { return consumeAngleOrNumberOrNoneRawAllowingSymbolTableIdent(args, symbolTable, context.mode); };
-    auto whitenessAndBlacknessConsumer = [&symbolTable](auto& args) { return consumePercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable); };
-    auto alphaConsumer = [&symbolTable](auto& args) { return consumeOptionalAlphaRawAllowingSymbolTableIdent(args, symbolTable); };
-
-    return parseHWBParametersRaw(args, WTFMove(hueConsumer), WTFMove(whitenessAndBlacknessConsumer), WTFMove(alphaConsumer));
-}
-
-static Color parseNonRelativeHWBParametersRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    auto hueConsumer = [&context](auto& args) { return consumeAngleOrNumberOrNoneRaw(args, context.mode); };
-    auto whitenessAndBlacknessConsumer = [](auto& args) { return consumePercentOrNoneRaw(args); };
-    auto alphaConsumer = [](auto& args) { return consumeOptionalAlphaRaw(args); };
-
-    return parseHWBParametersRaw(args, WTFMove(hueConsumer), WTFMove(whitenessAndBlacknessConsumer), WTFMove(alphaConsumer));
-}
-
-static Color parseHWBParametersRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    ASSERT(range.peek().functionId() == CSSValueHwb);
-
-    auto args = consumeFunction(range);
-
-    if (args.peek().id() == CSSValueFrom)
-        return parseRelativeHWBParametersRaw(args, context);
-    return parseNonRelativeHWBParametersRaw(args, context);
-}
-
-// The NormalizePercentage structs are specialized for the color types
-// that have specified rules for normalizing percentages by component.
-// The structs contains static members for the component that describe
-// the normalized value when the percent is 100%. As all treat 0% as
-// normalizing to 0, that is not encoded in the struct.
-template<typename ColorType> struct NormalizePercentage;
-
-template<> struct NormalizePercentage<Lab<float>> {
-    //  for L: 0% = 0.0, 100% = 100.0
-    //  for a and b: -100% = -125, 100% = 125 (NOTE: 0% is 0)
-
-    static constexpr double maximumLightnessNumber = 100.0;
-    static constexpr double lightnessScaleFactor = maximumLightnessNumber / 100.0;
-    static constexpr double abScaleFactor = 125.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<OKLab<float>> {
-    //  for L: 0% = 0.0, 100% = 1.0
-    //  for a and b: -100% = -0.4, 100% = 0.4 (NOTE: 0% is 0)
-
-    static constexpr double maximumLightnessNumber = 1.0;
-    static constexpr double lightnessScaleFactor = maximumLightnessNumber / 100.0;
-    static constexpr double abScaleFactor = 0.4 / 100.0;
-};
-
-template<> struct NormalizePercentage<LCHA<float>> {
-    //  for L: 0% = 0.0, 100% = 100.0
-    //  for C: 0% = 0, 100% = 150
-
-    static constexpr double maximumLightnessNumber = 100.0;
-    static constexpr double lightnessScaleFactor = maximumLightnessNumber / 100.0;
-    static constexpr double chromaScaleFactor = 150.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<OKLCHA<float>> {
-    //  for L: 0% = 0.0, 100% = 1.0
-    //  for C: 0% = 0.0 100% = 0.4
-
-    static constexpr double maximumLightnessNumber = 1.0;
-    static constexpr double lightnessScaleFactor = maximumLightnessNumber / 100.0;
-    static constexpr double chromaScaleFactor = 0.4 / 100.0;
-};
-
-template<> struct NormalizePercentage<XYZA<float, WhitePoint::D50>> {
-    //  for X,Y,Z: 0% = 0.0, 100% = 1.0
-
-    static constexpr double xyzScaleFactor = 1.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<XYZA<float, WhitePoint::D65>> {
-    //  for X,Y,Z: 0% = 0.0, 100% = 1.0
-
-    static constexpr double xyzScaleFactor = 1.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<ExtendedA98RGB<float>> {
-    //  for R,G,B: 0% = 0.0, 100% = 1.0
-
-    static constexpr double rgbScaleFactor = 1.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<ExtendedDisplayP3<float>> {
-    //  for R,G,B: 0% = 0.0, 100% = 1.0
-
-    static constexpr double rgbScaleFactor = 1.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<ExtendedProPhotoRGB<float>> {
-    //  for R,G,B: 0% = 0.0, 100% = 1.0
-
-    static constexpr double rgbScaleFactor = 1.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<ExtendedRec2020<float>> {
-    //  for R,G,B: 0% = 0.0, 100% = 1.0
-
-    static constexpr double rgbScaleFactor = 1.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<ExtendedSRGBA<float>> {
-    //  for R,G,B: 0% = 0.0, 100% = 1.0
-
-    static constexpr double rgbScaleFactor = 1.0 / 100.0;
-};
-
-template<> struct NormalizePercentage<ExtendedLinearSRGBA<float>> {
-    //  for R,G,B: 0% = 0.0, 100% = 1.0
-
-    static constexpr double rgbScaleFactor = 1.0 / 100.0;
-};
-
-template<typename ColorType>
-static double normalizeLightnessPercent(double percent)
-{
-    return NormalizePercentage<ColorType>::lightnessScaleFactor * percent;
-}
-
-template<typename ColorType>
-static double normalizeABPercent(double percent)
-{
-    return NormalizePercentage<ColorType>::abScaleFactor * percent;
-}
-
-template<typename ColorType>
-static double normalizeChromaPercent(double percent)
-{
-    return NormalizePercentage<ColorType>::chromaScaleFactor * percent;
-}
-
-template<typename ColorType>
-static double normalizeXYZPercent(double percent)
-{
-    return NormalizePercentage<ColorType>::xyzScaleFactor * percent;
-}
-
-template<typename ColorType>
-static double normalizeRGBPercent(double percent)
-{
-    return NormalizePercentage<ColorType>::rgbScaleFactor * percent;
-}
-
-template<typename ColorType, typename ConsumerForLightness, typename ConsumerForAB, typename ConsumerForAlpha>
-static Color parseLabParametersRaw(CSSParserTokenRange& args, ConsumerForLightness&& lightnessConsumer, ConsumerForAB&& abConsumer, ConsumerForAlpha&& alphaConsumer)
-{
-    auto lightness = lightnessConsumer(args);
-    if (!lightness)
-        return { };
-
-    auto aValue = abConsumer(args);
-    if (!aValue)
-        return { };
-
-    auto bValue = abConsumer(args);
-    if (!bValue)
-        return { };
-
-    auto alpha = alphaConsumer(args);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    auto normalizedLightness = WTF::switchOn(*lightness,
-        [] (NumberRaw number) { return std::clamp(number.value, 0.0, NormalizePercentage<ColorType>::maximumLightnessNumber); },
-        [] (PercentRaw percent) { return std::clamp(normalizeLightnessPercent<ColorType>(percent.value), 0.0, NormalizePercentage<ColorType>::maximumLightnessNumber); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-    auto normalizedA = WTF::switchOn(*aValue,
-        [] (NumberRaw number) { return number.value; },
-        [] (PercentRaw percent) { return normalizeABPercent<ColorType>(percent.value); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-    auto normalizedB = WTF::switchOn(*bValue,
-        [] (NumberRaw number) { return number.value; },
-        [] (PercentRaw percent) { return normalizeABPercent<ColorType>(percent.value); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-
-    return ColorType { static_cast<float>(normalizedLightness), static_cast<float>(normalizedA), static_cast<float>(normalizedB), static_cast<float>(*alpha) };
-}
-
-template<typename ColorType>
-static Color parseRelativeLabParametersRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    ASSERT(args.peek().id() == CSSValueFrom);
-    consumeIdentRaw(args);
-
-    auto originColor = consumeOriginColorRaw(args, context);
-    if (!originColor.isValid())
-        return { };
-
-    auto originColorAsLab = originColor.toColorTypeLossy<ColorType>().resolved();
-
-    CSSCalcSymbolTable symbolTable {
-        { CSSValueL, CSSUnitType::CSS_NUMBER, originColorAsLab.lightness },
-        { CSSValueA, CSSUnitType::CSS_NUMBER, originColorAsLab.a },
-        { CSSValueB, CSSUnitType::CSS_NUMBER, originColorAsLab.b },
-        { CSSValueAlpha, CSSUnitType::CSS_PERCENTAGE, originColorAsLab.alpha * 100.0 }
-    };
-
-    auto lightnessConsumer = [&symbolTable](auto& args) { return consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable); };
-    auto abConsumer = [&symbolTable](auto& args) { return consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable); };
-    auto alphaConsumer = [&symbolTable](auto& args) { return consumeOptionalAlphaRawAllowingSymbolTableIdent(args, symbolTable); };
-
-    return parseLabParametersRaw<ColorType>(args, WTFMove(lightnessConsumer), WTFMove(abConsumer), WTFMove(alphaConsumer));
-}
-
-template<typename ColorType>
-static Color parseNonRelativeLabParametersRaw(CSSParserTokenRange& args)
-{
-    auto lightnessConsumer = [](auto& args) { return consumeNumberOrPercentOrNoneRaw(args); };
-    auto abConsumer = [](auto& args) { return consumeNumberOrPercentOrNoneRaw(args); };
-    auto alphaConsumer = [](auto& args) { return consumeOptionalAlphaRaw(args); };
-
-    return parseLabParametersRaw<ColorType>(args, WTFMove(lightnessConsumer), WTFMove(abConsumer), WTFMove(alphaConsumer));
-}
-
-template<typename ColorType>
-static Color parseLabParametersRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    ASSERT(range.peek().functionId() == CSSValueLab || range.peek().functionId() == CSSValueOklab);
-
-    auto args = consumeFunction(range);
-
-    if (args.peek().id() == CSSValueFrom)
-        return parseRelativeLabParametersRaw<ColorType>(args, context);
-    return parseNonRelativeLabParametersRaw<ColorType>(args);
-}
-
-template<typename ColorType, typename ConsumerForLightness, typename ConsumerForChroma, typename ConsumerForHue, typename ConsumerForAlpha>
-static Color parseLCHParametersRaw(CSSParserTokenRange& args, ConsumerForLightness&& lightnessConsumer, ConsumerForChroma&& chromaConsumer, ConsumerForHue&& hueConsumer, ConsumerForAlpha&& alphaConsumer)
-{
-    auto lightness = lightnessConsumer(args);
-    if (!lightness)
-        return { };
-
-    auto chroma = chromaConsumer(args);
-    if (!chroma)
-        return { };
-
-    auto hue = hueConsumer(args);
-    if (!hue)
-        return { };
-
-    auto alpha = alphaConsumer(args);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    auto normalizedLightness = WTF::switchOn(*lightness,
-        [] (NumberRaw number) { return std::clamp(number.value, 0.0, NormalizePercentage<ColorType>::maximumLightnessNumber); },
-        [] (PercentRaw percent) { return std::clamp(normalizeLightnessPercent<ColorType>(percent.value), 0.0, NormalizePercentage<ColorType>::maximumLightnessNumber); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-    auto normalizedChroma = WTF::switchOn(*chroma,
-        [] (NumberRaw number) { return std::max(0.0, number.value); },
-        [] (PercentRaw percent) { return std::max(0.0, normalizeChromaPercent<ColorType>(percent.value)); },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-    auto normalizedHue = WTF::switchOn(*hue,
-        [] (AngleRaw angle) { return CSSPrimitiveValue::computeDegrees(angle.type, angle.value); },
-        [] (NumberRaw number) { return number.value; },
-        [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-    );
-
-    return ColorType { static_cast<float>(normalizedLightness), static_cast<float>(normalizedChroma), static_cast<float>(normalizedHue), static_cast<float>(*alpha) };
-}
-
-template<typename ColorType>
-static Color parseRelativeLCHParametersRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    ASSERT(args.peek().id() == CSSValueFrom);
-    consumeIdentRaw(args);
-
-    auto originColor = consumeOriginColorRaw(args, context);
-    if (!originColor.isValid())
-        return { };
-
-    auto originColorAsLCH = originColor.toColorTypeLossy<ColorType>().resolved();
-
-    CSSCalcSymbolTable symbolTable {
-        { CSSValueL, CSSUnitType::CSS_NUMBER, originColorAsLCH.lightness },
-        { CSSValueC, CSSUnitType::CSS_NUMBER, originColorAsLCH.chroma },
-        { CSSValueH, CSSUnitType::CSS_DEG, originColorAsLCH.hue },
-        { CSSValueAlpha, CSSUnitType::CSS_PERCENTAGE, originColorAsLCH.alpha * 100.0 }
-    };
-
-    auto lightnessConsumer = [&symbolTable](auto& args) { return consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable); };
-    auto chromaConsumer = [&symbolTable](auto& args) { return consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable); };
-    auto hueConsumer = [&symbolTable, &context](auto& args) { return consumeAngleOrNumberOrNoneRawAllowingSymbolTableIdent(args, symbolTable, context.mode); };
-    auto alphaConsumer = [&symbolTable](auto& args) { return consumeOptionalAlphaRawAllowingSymbolTableIdent(args, symbolTable); };
-
-    return parseLCHParametersRaw<ColorType>(args, WTFMove(lightnessConsumer), WTFMove(chromaConsumer), WTFMove(hueConsumer), WTFMove(alphaConsumer));
-}
-
-template<typename ColorType>
-static Color parseNonRelativeLCHParametersRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    auto lightnessConsumer = [](auto& args) { return consumeNumberOrPercentOrNoneRaw(args); };
-    auto chromaConsumer = [](auto& args) { return consumeNumberOrPercentOrNoneRaw(args); };
-    auto hueConsumer = [&context](auto& args) { return consumeAngleOrNumberOrNoneRaw(args, context.mode); };
-    auto alphaConsumer = [](auto& args) { return consumeOptionalAlphaRaw(args); };
-
-    return parseLCHParametersRaw<ColorType>(args, WTFMove(lightnessConsumer), WTFMove(chromaConsumer), WTFMove(hueConsumer), WTFMove(alphaConsumer));
-}
-
-template<typename ColorType>
-static Color parseLCHParametersRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    ASSERT(range.peek().functionId() == CSSValueLch || range.peek().functionId() == CSSValueOklch);
-
-    auto args = consumeFunction(range);
-
-    if (args.peek().id() == CSSValueFrom)
-        return parseRelativeLCHParametersRaw<ColorType>(args, context);
-    return parseNonRelativeLCHParametersRaw<ColorType>(args, context);
-}
-
-template<typename ColorType, typename ConsumerForRGB, typename ConsumerForAlpha>
-static Color parseColorFunctionForRGBTypesRaw(CSSParserTokenRange& args, ConsumerForRGB&& rgbConsumer, ConsumerForAlpha&& alphaConsumer)
-{
-    double channels[3] = { 0, 0, 0 };
-    for (auto& channel : channels) {
-        auto value = rgbConsumer(args);
-        if (!value)
-            return { };
-
-        channel = WTF::switchOn(*value,
-            [] (NumberRaw number) { return number.value; },
-            [] (PercentRaw percent) { return normalizeRGBPercent<ColorType>(percent.value); },
-            [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-        );
-    }
-
-    auto alpha = alphaConsumer(args);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    return { ColorType { static_cast<float>(channels[0]), static_cast<float>(channels[1]), static_cast<float>(channels[2]), static_cast<float>(*alpha) }, Color::Flags::UseColorFunctionSerialization };
-}
-
-template<typename ColorType> static Color parseRelativeColorFunctionForRGBTypes(CSSParserTokenRange& args, Color originColor)
-{
-    ASSERT(args.peek().id() == CSSValueA98Rgb || args.peek().id() == CSSValueDisplayP3 || args.peek().id() == CSSValueProphotoRgb || args.peek().id() == CSSValueRec2020 || args.peek().id() == CSSValueSRGB || args.peek().id() == CSSValueSrgbLinear);
-
-    consumeIdentRaw(args);
-
-    auto originColorAsColorType = originColor.toColorTypeLossy<ColorType>().resolved();
-
-    CSSCalcSymbolTable symbolTable {
-        { CSSValueR, CSSUnitType::CSS_PERCENTAGE, originColorAsColorType.red * 100.0 },
-        { CSSValueG, CSSUnitType::CSS_PERCENTAGE, originColorAsColorType.green * 100.0 },
-        { CSSValueB, CSSUnitType::CSS_PERCENTAGE, originColorAsColorType.blue * 100.0 },
-        { CSSValueAlpha, CSSUnitType::CSS_PERCENTAGE, originColorAsColorType.alpha * 100.0 }
-    };
-
-    auto consumeRGB = [&symbolTable](auto& args) { return consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable, ValueRange::All); };
-    auto consumeAlpha = [&symbolTable](auto& args) { return consumeOptionalAlphaRawAllowingSymbolTableIdent(args, symbolTable); };
-
-    return parseColorFunctionForRGBTypesRaw<ColorType>(args, WTFMove(consumeRGB), WTFMove(consumeAlpha));
-}
-
-template<typename ColorType> static Color parseColorFunctionForRGBTypesRaw(CSSParserTokenRange& args)
-{
-    ASSERT(args.peek().id() == CSSValueA98Rgb || args.peek().id() == CSSValueDisplayP3 || args.peek().id() == CSSValueProphotoRgb || args.peek().id() == CSSValueRec2020 || args.peek().id() == CSSValueSRGB || args.peek().id() == CSSValueSrgbLinear);
-
-    consumeIdentRaw(args);
-
-    auto consumeRGB = [](auto& args) { return consumeNumberOrPercentOrNoneRaw(args); };
-    auto consumeAlpha = [](auto& args) { return consumeOptionalAlphaRaw(args); };
-
-    return parseColorFunctionForRGBTypesRaw<ColorType>(args, WTFMove(consumeRGB), WTFMove(consumeAlpha));
-}
-
-template<typename ColorType, typename ConsumerForXYZ, typename ConsumerForAlpha>
-static Color parseColorFunctionForXYZTypesRaw(CSSParserTokenRange& args, ConsumerForXYZ&& xyzConsumer, ConsumerForAlpha&& alphaConsumer)
-{
-    double channels[3] = { 0, 0, 0 };
-    for (auto& channel : channels) {
-        auto value = xyzConsumer(args);
-        if (!value)
-            return { };
-
-        channel = WTF::switchOn(*value,
-            [] (NumberRaw number) { return number.value; },
-            [] (PercentRaw percent) { return normalizeXYZPercent<ColorType>(percent.value); },
-            [] (NoneRaw) { return std::numeric_limits<double>::quiet_NaN(); }
-        );
-    }
-
-    auto alpha = alphaConsumer(args);
-    if (!alpha)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    return { ColorType { static_cast<float>(channels[0]), static_cast<float>(channels[1]), static_cast<float>(channels[2]), static_cast<float>(*alpha) }, Color::Flags::UseColorFunctionSerialization };
-}
-
-template<typename ColorType> static Color parseRelativeColorFunctionForXYZTypes(CSSParserTokenRange& args, Color originColor)
-{
-    ASSERT(args.peek().id() == CSSValueXyz || args.peek().id() == CSSValueXyzD50 || args.peek().id() == CSSValueXyzD65);
-
-    consumeIdentRaw(args);
-
-    auto originColorAsXYZ = originColor.toColorTypeLossy<ColorType>().resolved();
-
-    CSSCalcSymbolTable symbolTable {
-        { CSSValueX, CSSUnitType::CSS_NUMBER, originColorAsXYZ.x },
-        { CSSValueY, CSSUnitType::CSS_NUMBER, originColorAsXYZ.y },
-        { CSSValueZ, CSSUnitType::CSS_NUMBER, originColorAsXYZ.z },
-        { CSSValueAlpha, CSSUnitType::CSS_PERCENTAGE, originColorAsXYZ.alpha * 100.0 }
-    };
-
-    auto consumeXYZ = [&symbolTable](auto& args) { return consumeNumberOrPercentOrNoneRawAllowingSymbolTableIdent(args, symbolTable); };
-    auto consumeAlpha = [&symbolTable](auto& args) { return consumeOptionalAlphaRawAllowingSymbolTableIdent(args, symbolTable); };
-
-    return parseColorFunctionForXYZTypesRaw<ColorType>(args, WTFMove(consumeXYZ), WTFMove(consumeAlpha));
-}
-
-template<typename ColorType> static Color parseColorFunctionForXYZTypesRaw(CSSParserTokenRange& args)
-{
-    ASSERT(args.peek().id() == CSSValueXyz || args.peek().id() == CSSValueXyzD50 || args.peek().id() == CSSValueXyzD65);
-
-    consumeIdentRaw(args);
-
-    auto consumeXYZ = [](auto& args) { return consumeNumberOrPercentOrNoneRaw(args); };
-    auto consumeAlpha = [](auto& args) { return consumeOptionalAlphaRaw(args); };
-
-    return parseColorFunctionForXYZTypesRaw<ColorType>(args, WTFMove(consumeXYZ), WTFMove(consumeAlpha));
-}
-
-static Color parseRelativeColorFunctionParameters(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    ASSERT(args.peek().id() == CSSValueFrom);
-    consumeIdentRaw(args);
-
-    auto originColor = consumeOriginColorRaw(args, context);
-    if (!originColor.isValid())
-        return { };
-
-    switch (args.peek().id()) {
-    case CSSValueA98Rgb:
-        return parseRelativeColorFunctionForRGBTypes<ExtendedA98RGB<float>>(args, WTFMove(originColor));
-    case CSSValueDisplayP3:
-        return parseRelativeColorFunctionForRGBTypes<ExtendedDisplayP3<float>>(args, WTFMove(originColor));
-    case CSSValueProphotoRgb:
-        return parseRelativeColorFunctionForRGBTypes<ExtendedProPhotoRGB<float>>(args, WTFMove(originColor));
-    case CSSValueRec2020:
-        return parseRelativeColorFunctionForRGBTypes<ExtendedRec2020<float>>(args, WTFMove(originColor));
-    case CSSValueSRGB:
-        return parseRelativeColorFunctionForRGBTypes<ExtendedSRGBA<float>>(args, WTFMove(originColor));
-    case CSSValueSrgbLinear:
-        return parseRelativeColorFunctionForRGBTypes<ExtendedLinearSRGBA<float>>(args, WTFMove(originColor));
-    case CSSValueXyzD50:
-        return parseRelativeColorFunctionForXYZTypes<XYZA<float, WhitePoint::D50>>(args, WTFMove(originColor));
-    case CSSValueXyz:
-    case CSSValueXyzD65:
-        return parseRelativeColorFunctionForXYZTypes<XYZA<float, WhitePoint::D65>>(args, WTFMove(originColor));
-    default:
-        return { };
-    }
-
-    ASSERT_NOT_REACHED();
-    return { };
-}
-
-static Color parseNonRelativeColorFunctionParameters(CSSParserTokenRange& args)
-{
-    switch (args.peek().id()) {
-    case CSSValueA98Rgb:
-        return parseColorFunctionForRGBTypesRaw<ExtendedA98RGB<float>>(args);
-    case CSSValueDisplayP3:
-        return parseColorFunctionForRGBTypesRaw<ExtendedDisplayP3<float>>(args);
-    case CSSValueProphotoRgb:
-        return parseColorFunctionForRGBTypesRaw<ExtendedProPhotoRGB<float>>(args);
-    case CSSValueRec2020:
-        return parseColorFunctionForRGBTypesRaw<ExtendedRec2020<float>>(args);
-    case CSSValueSRGB:
-        return parseColorFunctionForRGBTypesRaw<ExtendedSRGBA<float>>(args);
-    case CSSValueSrgbLinear:
-        return parseColorFunctionForRGBTypesRaw<ExtendedLinearSRGBA<float>>(args);
-    case CSSValueXyzD50:
-        return parseColorFunctionForXYZTypesRaw<XYZA<float, WhitePoint::D50>>(args);
-    case CSSValueXyz:
-    case CSSValueXyzD65:
-        return parseColorFunctionForXYZTypesRaw<XYZA<float, WhitePoint::D65>>(args);
-    default:
-        return { };
-    }
-
-    ASSERT_NOT_REACHED();
-    return { };
-}
-
-static Color parseColorFunctionParametersRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    ASSERT(range.peek().functionId() == CSSValueColor);
-    auto args = consumeFunction(range);
-
-    auto color = [&] {
-        if (args.peek().id() == CSSValueFrom)
-            return parseRelativeColorFunctionParameters(args, context);
-        return parseNonRelativeColorFunctionParameters(args);
-    }();
-
-    ASSERT(!color.isValid() || color.usesColorFunctionSerialization());
-    return color;
-}
-
-static Color selectFirstColorThatMeetsOrExceedsTargetContrast(const Color& originBackgroundColor, Vector<Color>&& colorsToCompareAgainst, double targetContrast)
-{
-    auto originBackgroundColorLuminance = originBackgroundColor.luminance();
-
-    for (auto& color : colorsToCompareAgainst) {
-        if (contrastRatio(originBackgroundColorLuminance, color.luminance()) >= targetContrast)
-            return WTFMove(color);
-    }
-
-    // If there is a target contrast, and the end of the list is reached without meeting that target,
-    // either white or black is returned, whichever has the higher contrast.
-    auto contrastWithWhite = contrastRatio(originBackgroundColorLuminance, 1.0);
-    auto contrastWithBlack = contrastRatio(originBackgroundColorLuminance, 0.0);
-    return contrastWithWhite > contrastWithBlack ? Color::white : Color::black;
-}
-
-static Color selectFirstColorWithHighestContrast(const Color& originBackgroundColor, Vector<Color>&& colorsToCompareAgainst)
-{
-    auto originBackgroundColorLuminance = originBackgroundColor.luminance();
-
-    auto* colorWithGreatestContrast = &colorsToCompareAgainst[0];
-    double greatestContrastSoFar = 0;
-    for (auto& color : colorsToCompareAgainst) {
-        auto contrast = contrastRatio(originBackgroundColorLuminance, color.luminance());
-        if (contrast > greatestContrastSoFar) {
-            greatestContrastSoFar = contrast;
-            colorWithGreatestContrast = &color;
-        }
-    }
-
-    return WTFMove(*colorWithGreatestContrast);
-}
-
-static Color parseColorContrastFunctionParametersRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    ASSERT(range.peek().functionId() == CSSValueColorContrast);
-
-    if (!context.colorContrastEnabled)
-        return { };
-
-    auto args = consumeFunction(range);
-
-    auto originBackgroundColor = consumeOriginColorRaw(args, context);
-    if (!originBackgroundColor.isValid())
-        return { };
-
-    if (!consumeIdentRaw<CSSValueVs>(args))
-        return { };
-
-    Vector<Color> colorsToCompareAgainst;
-    bool consumedTo = false;
-    do {
-        auto colorToCompareAgainst = consumeOriginColorRaw(args, context);
-        if (!colorToCompareAgainst.isValid())
-            return { };
-
-        colorsToCompareAgainst.append(WTFMove(colorToCompareAgainst));
-
-        if (consumeIdentRaw<CSSValueTo>(args)) {
-            consumedTo = true;
-            break;
-        }
-    } while (consumeCommaIncludingWhitespace(args));
-
-    // Handle the invalid case where there is only one color in the "compare against" color list.
-    if (colorsToCompareAgainst.size() == 1)
-        return { };
-
-    if (consumedTo) {
-        auto targetContrast = [&] () -> std::optional<NumberRaw> {
-            if (args.peek().type() == IdentToken) {
-                static constexpr std::pair<CSSValueID, NumberRaw> targetContrastMappings[] {
-                    { CSSValueAA, NumberRaw { 4.5 } },
-                    { CSSValueAALarge, NumberRaw { 3.0 } },
-                    { CSSValueAAA, NumberRaw { 7.0 } },
-                    { CSSValueAAALarge, NumberRaw { 4.5 } },
-                };
-                static constexpr SortedArrayMap targetContrastMap { targetContrastMappings };
-                auto value = targetContrastMap.tryGet(args.consumeIncludingWhitespace().id());
-                return value ? std::make_optional(*value) : std::nullopt;
-            }
-            return consumeNumberRaw(args);
-        }();
-
-        if (!targetContrast)
-            return { };
-
-        // Handle the invalid case where there are additional tokens after the target contrast.
-        if (!args.atEnd())
-            return { };
-
-        // When a target constast is specified, we select "the first color color to meet or exceed the target contrast."
-        return selectFirstColorThatMeetsOrExceedsTargetContrast(originBackgroundColor, WTFMove(colorsToCompareAgainst), targetContrast->value);
-    }
-
-    // Handle the invalid case where there are additional tokens after the "compare against" color list that are not the 'to' identifier.
-    if (!args.atEnd())
-        return { };
-
-    // When a target constast is NOT specified, we select "the first color with the highest contrast to the single color."
-    return selectFirstColorWithHighestContrast(originBackgroundColor, WTFMove(colorsToCompareAgainst));
-}
-
-static std::optional<HueInterpolationMethod> consumeHueInterpolationMethod(CSSParserTokenRange& range)
-{
-    static constexpr std::pair<CSSValueID, HueInterpolationMethod> hueInterpolationMethodMappings[] {
-        { CSSValueShorter, HueInterpolationMethod::Shorter },
-        { CSSValueLonger, HueInterpolationMethod::Longer },
-        { CSSValueIncreasing, HueInterpolationMethod::Increasing },
-        { CSSValueDecreasing, HueInterpolationMethod::Decreasing },
-    };
-    static constexpr SortedArrayMap hueInterpolationMethodMap { hueInterpolationMethodMappings };
-
-    return consumeIdentUsingMapping(range, hueInterpolationMethodMap);
-}
-
-static std::optional<ColorInterpolationMethod> consumeColorInterpolationMethod(CSSParserTokenRange& args)
-{
-    // <rectangular-color-space> = srgb | srgb-linear | display-p3 | a98-rgb | prophoto-rgb | rec2020 | lab | oklab | xyz | xyz-d50 | xyz-d65
-    // <polar-color-space> = hsl | hwb | lch | oklch
-    // <hue-interpolation-method> = [ shorter | longer | increasing | decreasing ] hue
-    // <color-interpolation-method> = in [ <rectangular-color-space> | <polar-color-space> <hue-interpolation-method>? ]
-
-    ASSERT(args.peek().id() == CSSValueIn);
-    consumeIdentRaw(args);
-
-    auto consumePolarColorSpace = [](CSSParserTokenRange& args, auto colorInterpolationMethod) -> std::optional<ColorInterpolationMethod> {
-        // Consume the color space identifier.
-        args.consumeIncludingWhitespace();
-
-        // <hue-interpolation-method> is optional, so if it is not provided, we just use the default value
-        // specified in the passed in 'colorInterpolationMethod' parameter.
-        auto hueInterpolationMethod = consumeHueInterpolationMethod(args);
-        if (!hueInterpolationMethod)
-            return {{ colorInterpolationMethod, AlphaPremultiplication::Premultiplied }};
-
-        // If the hue-interpolation-method was provided it must be followed immediately by the 'hue' identifier.
-        if (!consumeIdentRaw<CSSValueHue>(args))
-            return { };
-
-        colorInterpolationMethod.hueInterpolationMethod = *hueInterpolationMethod;
-
-        return {{ colorInterpolationMethod, AlphaPremultiplication::Premultiplied }};
-    };
-
-    auto consumeRectangularColorSpace = [](CSSParserTokenRange& args, auto colorInterpolationMethod) -> std::optional<ColorInterpolationMethod> {
-        // Consume the color space identifier.
-        args.consumeIncludingWhitespace();
-
-        return {{ colorInterpolationMethod, AlphaPremultiplication::Premultiplied }};
-    };
-
-    switch (args.peek().id()) {
-    case CSSValueHsl:
-        return consumePolarColorSpace(args, ColorInterpolationMethod::HSL { });
-    case CSSValueHwb:
-        return consumePolarColorSpace(args, ColorInterpolationMethod::HWB { });
-    case CSSValueLch:
-        return consumePolarColorSpace(args, ColorInterpolationMethod::LCH { });
-    case CSSValueLab:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::Lab { });
-    case CSSValueOklch:
-        return consumePolarColorSpace(args, ColorInterpolationMethod::OKLCH { });
-    case CSSValueOklab:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::OKLab { });
-    case CSSValueSRGB:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::SRGB { });
-    case CSSValueSrgbLinear:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::SRGBLinear { });
-    case CSSValueDisplayP3:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::DisplayP3 { });
-    case CSSValueA98Rgb:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::A98RGB { });
-    case CSSValueProphotoRgb:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::ProPhotoRGB { });
-    case CSSValueRec2020:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::Rec2020 { });
-    case CSSValueXyzD50:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::XYZD50 { });
-    case CSSValueXyz:
-    case CSSValueXyzD65:
-        return consumeRectangularColorSpace(args, ColorInterpolationMethod::XYZD65 { });
-    default:
-        return { };
-    }
-}
-
-static std::optional<CSSResolvedColorMix::Component> consumeColorMixComponentRaw(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    CSSResolvedColorMix::Component result;
-
-    if (auto percentage = consumePercentRaw(args)) {
-        if (percentage->value < 0.0 || percentage->value > 100.0)
-            return { };
-        result.percentage = percentage->value;
-    }
-
-    result.color = consumeOriginColorRaw(args, context);
-    if (!result.color.isValid())
-        return std::nullopt;
-
-    if (!result.percentage) {
-        if (auto percentage = consumePercentRaw(args)) {
-            if (percentage->value < 0.0 || percentage->value > 100.0)
-                return { };
-            result.percentage = percentage->value;
-        }
-    }
-
-    return result;
-}
-
-static std::optional<CSSUnresolvedColorMix::Component> consumeColorMixComponent(CSSParserTokenRange& args, const CSSParserContext& context)
-{
-    // [ <color> && <percentage [0,100]>? ]
-
-    RefPtr<CSSPrimitiveValue> color;
-    RefPtr<CSSPrimitiveValue> percentage;
-
-    if (auto percent = consumePercent(args, ValueRange::All)) {
-        if (!percent->isCalculated()) {
-            auto value = percent->doubleValue();
-            if (value < 0.0 || value > 100.0)
-                return std::nullopt;
-        }
-        percentage = percent;
-    }
-
-    auto originColor = consumeColor(args, context);
-    if (!originColor)
-        return std::nullopt;
-
-    if (!percentage) {
-        if (auto percent = consumePercent(args, ValueRange::All)) {
-            if (!percent->isCalculated()) {
-                auto value = percent->doubleValue();
-                if (value < 0.0 || value > 100.0)
-                    return std::nullopt;
-            }
-            percentage = percent;
-        }
-    }
-
-    return CSSUnresolvedColorMix::Component {
-        originColor.releaseNonNull(),
-        WTFMove(percentage)
-    };
-}
-
-static Color parseColorMixFunctionParametersRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    // color-mix() = color-mix( <color-interpolation-method> , [ <color> && <percentage [0,100]>? ]#{2})
-
-    ASSERT(range.peek().functionId() == CSSValueColorMix);
-
-    auto args = consumeFunction(range);
-
-    if (args.peek().id() != CSSValueIn)
-        return { };
-
-    auto colorInterpolationMethod = consumeColorInterpolationMethod(args);
-    if (!colorInterpolationMethod)
-        return { };
-
-    if (!consumeCommaIncludingWhitespace(args))
-        return { };
-
-    auto mixComponent1 = consumeColorMixComponentRaw(args, context);
-    if (!mixComponent1)
-        return { };
-
-    if (!consumeCommaIncludingWhitespace(args))
-        return { };
-
-    auto mixComponent2 = consumeColorMixComponentRaw(args, context);
-    if (!mixComponent2)
-        return { };
-
-    if (!args.atEnd())
-        return { };
-
-    return mix(CSSResolvedColorMix {
-        WTFMove(*colorInterpolationMethod),
-        WTFMove(*mixComponent1),
-        WTFMove(*mixComponent2)
-    });
-}
-
-static bool hasNonCalculatedZeroPercentage(const CSSUnresolvedColorMix::Component& mixComponent)
-{
-    return mixComponent.percentage && !mixComponent.percentage->isCalculated() && mixComponent.percentage->doubleValue() == 0.0;
-}
-
-static std::optional<ColorOrUnresolvedColor> parseColorMixFunctionParameters(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    // color-mix() = color-mix( <color-interpolation-method> , [ <color> && <percentage [0,100]>? ]#{2})
-
-    ASSERT(range.peek().functionId() == CSSValueColorMix);
-
-    auto args = consumeFunction(range);
-
-    if (args.peek().id() != CSSValueIn)
-        return std::nullopt;
-
-    auto colorInterpolationMethod = consumeColorInterpolationMethod(args);
-    if (!colorInterpolationMethod)
-        return std::nullopt;
-
-    if (!consumeCommaIncludingWhitespace(args))
-        return std::nullopt;
-
-    auto mixComponent1 = consumeColorMixComponent(args, context);
-    if (!mixComponent1)
-        return std::nullopt;
-
-    if (!consumeCommaIncludingWhitespace(args))
-        return std::nullopt;
-
-    auto mixComponent2 = consumeColorMixComponent(args, context);
-    if (!mixComponent2)
-        return std::nullopt;
-
-    if (!args.atEnd())
-        return std::nullopt;
-
-    if (hasNonCalculatedZeroPercentage(*mixComponent1) && hasNonCalculatedZeroPercentage(*mixComponent2)) {
-        // This eagerly marks the parse as invalid if both percentage components are non-calc
-        // and equal to 0. This satisfies step 5 of section 2.1. Percentage Normalization in
-        // https://w3c.github.io/csswg-drafts/css-color-5/#color-mix which states:
-        //    "If the percentages sum to zero, the function is invalid."
-        //
-        // The only way the percentages can sum to zero is both are zero, since we reject any
-        // percentages less than zero, and missing percentages are treated as "100% - p(other)".
-        //
-        // FIXME: Should we also do this for calculated values? Or should we let the parse be
-        // valid and fail to produce a valid color at style building time.
-        return std::nullopt;
-    }
-
-    return { CSSUnresolvedColor { CSSUnresolvedColorMix {
-        WTFMove(*colorInterpolationMethod),
-        WTFMove(*mixComponent1),
-        WTFMove(*mixComponent2)
-    } } };
-}
-
-static std::optional<ColorOrUnresolvedColor> parseLightDarkFunctionParameters(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    // light-dark() = light-dark( <color>, <color> )
-
-    ASSERT(range.peek().functionId() == CSSValueLightDark);
-
-    if (!context.lightDarkEnabled)
-        return std::nullopt;
-
-    auto args = consumeFunction(range);
-
-    auto lightColor = consumeColor(args, context);
-    if (!lightColor)
-        return std::nullopt;
-
-    if (!consumeCommaIncludingWhitespace(args))
-        return std::nullopt;
-
-    auto darkColor = consumeColor(args, context);
-    if (!darkColor)
-        return std::nullopt;
-
-    if (!args.atEnd())
-        return std::nullopt;
-
-    return { CSSUnresolvedColor { CSSUnresolvedLightDark {
-        lightColor.releaseNonNull(),
-        darkColor.releaseNonNull()
-    } } };
-}
-
-static std::optional<SRGBA<uint8_t>> parseHexColor(CSSParserTokenRange& range, bool acceptQuirkyColors)
-{
-    String string;
-    StringView view;
-    auto& token = range.peek();
-    if (token.type() == HashToken)
-        view = token.value();
-    else {
-        if (!acceptQuirkyColors)
-            return std::nullopt;
-        if (token.type() == IdentToken) {
-            view = token.value(); // e.g. FF0000
-            if (view.length() != 3 && view.length() != 6)
-                return std::nullopt;
-        } else if (token.type() == NumberToken || token.type() == DimensionToken) {
-            if (token.numericValueType() != IntegerValueType)
-                return std::nullopt;
-            auto numericValue = token.numericValue();
-            if (!(numericValue >= 0 && numericValue < 1000000))
-                return std::nullopt;
-            auto integerValue = static_cast<int>(token.numericValue());
-            if (token.type() == NumberToken)
-                string = String::number(integerValue); // e.g. 112233
-            else
-                string = makeString(integerValue, token.unitString()); // e.g. 0001FF
-            if (string.length() < 6)
-                string = makeString(&"000000"[string.length()], string);
-
-            if (string.length() != 3 && string.length() != 6)
-                return std::nullopt;
-            view = string;
-        } else
-            return std::nullopt;
-    }
-    auto result = CSSParser::parseHexColor(view);
-    if (!result)
-        return std::nullopt;
-    range.consumeIncludingWhitespace();
-    return *result;
-}
-
-static Color parseColorFunctionRaw(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    CSSParserTokenRange colorRange = range;
-    CSSValueID functionId = range.peek().functionId();
-    Color color;
-    switch (functionId) {
-    case CSSValueRgb:
-        color = parseRGBParametersRaw<RGBFunctionMode::RGB>(colorRange, context);
-        break;
-    case CSSValueRgba:
-        color = parseRGBParametersRaw<RGBFunctionMode::RGBA>(colorRange, context);
-        break;
-    case CSSValueHsl:
-        color = parseHSLParametersRaw<HSLFunctionMode::HSL>(colorRange, context);
-        break;
-    case CSSValueHsla:
-        color = parseHSLParametersRaw<HSLFunctionMode::HSLA>(colorRange, context);
-        break;
-    case CSSValueHwb:
-        color = parseHWBParametersRaw(colorRange, context);
-        break;
-    case CSSValueLab:
-        color = parseLabParametersRaw<Lab<float>>(colorRange, context);
-        break;
-    case CSSValueLch:
-        color = parseLCHParametersRaw<LCHA<float>>(colorRange, context);
-        break;
-    case CSSValueOklab:
-        color = parseLabParametersRaw<OKLab<float>>(colorRange, context);
-        break;
-    case CSSValueOklch:
-        color = parseLCHParametersRaw<OKLCHA<float>>(colorRange, context);
-        break;
-    case CSSValueColor:
-        color = parseColorFunctionParametersRaw(colorRange, context);
-        break;
-    case CSSValueColorContrast:
-        color = parseColorContrastFunctionParametersRaw(colorRange, context);
-        break;
-    case CSSValueColorMix:
-        color = parseColorMixFunctionParametersRaw(colorRange, context);
-        break;
-    case CSSValueLightDark:
-        // FIXME: Need a worker-safe way to compute light-dark colors.
-        return { };
-    default:
-        return { };
-    }
-    if (color.isValid())
-        range = colorRange;
-    return color;
-}
-
-static std::optional<ColorOrUnresolvedColor> parseColorFunction(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    auto checkColor = [] (Color&& color) -> std::optional<ColorOrUnresolvedColor> {
-        if (!color.isValid())
-            return std::nullopt;
-        return ColorOrUnresolvedColor { WTFMove(color) };
-    };
-
-    CSSParserTokenRange colorRange = range;
-    CSSValueID functionId = range.peek().functionId();
-    std::optional<ColorOrUnresolvedColor> color;
-    switch (functionId) {
-    case CSSValueRgb:
-        color = checkColor(parseRGBParametersRaw<RGBFunctionMode::RGB>(colorRange, context));
-        break;
-    case CSSValueRgba:
-        color = checkColor(parseRGBParametersRaw<RGBFunctionMode::RGBA>(colorRange, context));
-        break;
-    case CSSValueHsl:
-        color = checkColor(parseHSLParametersRaw<HSLFunctionMode::HSL>(colorRange, context));
-        break;
-    case CSSValueHsla:
-        color = checkColor(parseHSLParametersRaw<HSLFunctionMode::HSLA>(colorRange, context));
-        break;
-    case CSSValueHwb:
-        color = checkColor(parseHWBParametersRaw(colorRange, context));
-        break;
-    case CSSValueLab:
-        color = checkColor(parseLabParametersRaw<Lab<float>>(colorRange, context));
-        break;
-    case CSSValueLch:
-        color = checkColor(parseLCHParametersRaw<LCHA<float>>(colorRange, context));
-        break;
-    case CSSValueOklab:
-        color = checkColor(parseLabParametersRaw<OKLab<float>>(colorRange, context));
-        break;
-    case CSSValueOklch:
-        color = checkColor(parseLCHParametersRaw<OKLCHA<float>>(colorRange, context));
-        break;
-    case CSSValueColor:
-        color = checkColor(parseColorFunctionParametersRaw(colorRange, context));
-        break;
-    case CSSValueColorContrast:
-        color = checkColor(parseColorContrastFunctionParametersRaw(colorRange, context));
-        break;
-    case CSSValueColorMix:
-        color = parseColorMixFunctionParameters(colorRange, context);
-        break;
-    case CSSValueLightDark:
-        color = parseLightDarkFunctionParameters(colorRange, context);
-        break;
-    default:
-        return { };
-    }
-    if (color)
-        range = colorRange;
-    return color;
-}
-
-Color consumeColorWorkerSafe(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    Color result;
-    auto keyword = range.peek().id();
-    if (StyleColor::isColorKeyword(keyword)) {
-        // FIXME: Need a worker-safe way to compute the system colors.
-        //        For now, we detect the system color, but then intentionally fail parsing.
-        if (StyleColor::isSystemColorKeyword(keyword))
-            return { };
-        if (!isColorKeywordAllowedInMode(keyword, context.mode))
-            return { };
-        result = StyleColor::colorFromKeyword(keyword, { });
-        range.consumeIncludingWhitespace();
-    }
-
-    if (auto parsedColor = parseHexColor(range, false))
-        result = *parsedColor;
-    else
-        result = parseColorFunctionRaw(range, context);
-
-    if (!range.atEnd())
-        return { };
-
-    return result;
-}
-
-RefPtr<CSSPrimitiveValue> consumeColor(CSSParserTokenRange& range, const CSSParserContext& context, bool acceptQuirkyColors, OptionSet<StyleColor::CSSColorType> allowedColorTypes)
-{
-    auto keyword = range.peek().id();
-    if (StyleColor::isColorKeyword(keyword, allowedColorTypes)) {
-        if (!isColorKeywordAllowedInMode(keyword, context.mode))
-            return nullptr;
-        return consumeIdent(range);
-    }
-
-    if (auto parsedColor = parseHexColor(range, acceptQuirkyColors))
-        return CSSValuePool::singleton().createColorValue(Color { *parsedColor });
-
-    if (auto colorOrUnresolvedColor = parseColorFunction(range, context)) {
-        return WTF::switchOn(WTFMove(*colorOrUnresolvedColor),
-            [] (Color&& color) -> RefPtr<CSSPrimitiveValue> {
-                return CSSValuePool::singleton().createColorValue(WTFMove(color));
-            },
-            [] (CSSUnresolvedColor&& color) -> RefPtr<CSSPrimitiveValue> {
-                return CSSPrimitiveValue::create(WTFMove(color));
-            }
-        );
-    }
-
-    return nullptr;
 }
 
 static RefPtr<CSSPrimitiveValue> consumePositionComponent(CSSParserTokenRange& range, CSSParserMode parserMode, UnitlessQuirk unitless, NegativePercentagePolicy negativePercentagePolicy = NegativePercentagePolicy::Forbid)
